+++ Assets/UnitySensors/Tests/Editor/GeoCoordinateBasicTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7242262bdab76d8c8c4924acb5f99f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
