+++ Assets/UnitySensors/Tests/Editor/HeaderSerializerTests.cs
@@ -1,0 +1,218 @@
+using NUnit.Framework;
+using UnityEngine;
+
+namespace UnitySensors.Tests.Editor
+{
+    [TestFixture]
+    public class HeaderSerializerTests
+    {
+        [Test]
+        public void HeaderSerializer_ReflectionAccess_ShouldBeAccessible()
+        {
+            // Test that HeaderSerializer can be accessed via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    Assert.IsTrue(type.IsClass);
+                    Assert.IsFalse(type.IsAbstract);
+                    Assert.IsTrue(type.IsPublic);
+                    
+                    // Check System.Serializable attribute
+                    var attrs = type.GetCustomAttributes(typeof(System.SerializableAttribute), false);
+                    Assert.Greater(attrs.Length, 0, "Should have System.Serializable attribute");
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_Inheritance_ShouldExtendRosMsgSerializer()
+        {
+            // Test that HeaderSerializer properly inherits from RosMsgSerializer<HeaderMsg>
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    var baseType = type.BaseType;
+                    Assert.IsNotNull(baseType);
+                    Assert.IsTrue(baseType.IsGenericType);
+                    
+                    // Check that it's RosMsgSerializer<HeaderMsg>
+                    var genericTypeDef = baseType.GetGenericTypeDefinition();
+                    var expectedBase = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
+                    if (expectedBase != null)
+                    {
+                        Assert.AreEqual(expectedBase, genericTypeDef);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_Fields_ShouldBeAccessible()
+        {
+            // Test that HeaderSerializer fields are properly defined
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    // Check _source field
+                    var sourceField = type.GetField("_source", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (sourceField != null)
+                    {
+                        Assert.IsNotNull(sourceField);
+                        Assert.AreEqual(typeof(UnityEngine.Object), sourceField.FieldType);
+                    }
+                    
+                    // Check _frame_id field
+                    var frameIdField = type.GetField("_frame_id", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (frameIdField != null)
+                    {
+                        Assert.IsNotNull(frameIdField);
+                        Assert.AreEqual(typeof(string), frameIdField.FieldType);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_InitMethod_ShouldBeOverridden()
+        {
+            // Test that Init method is properly overridden
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    var initMethod = type.GetMethod("Init");
+                    if (initMethod != null)
+                    {
+                        Assert.IsNotNull(initMethod);
+                        Assert.IsTrue(initMethod.IsPublic);
+                        Assert.IsTrue(initMethod.IsVirtual);
+                        Assert.IsFalse(initMethod.IsAbstract);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_SerializeMethod_ShouldBeOverridden()
+        {
+            // Test that Serialize method is properly overridden
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    var serializeMethod = type.GetMethod("Serialize");
+                    if (serializeMethod != null)
+                    {
+                        Assert.IsNotNull(serializeMethod);
+                        Assert.IsTrue(serializeMethod.IsPublic);
+                        Assert.IsTrue(serializeMethod.IsVirtual);
+                        Assert.IsFalse(serializeMethod.IsAbstract);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_ITimeInterface_ShouldBeAccessible()
+        {
+            // Test that ITimeInterface is properly used
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var timeInterfaceType = System.Type.GetType("UnitySensors.Interface.Std.ITimeInterface, UnitySensorsRuntime");
+                if (timeInterfaceType != null)
+                {
+                    Assert.IsTrue(timeInterfaceType.IsInterface);
+                    Assert.IsTrue(timeInterfaceType.IsPublic);
+                    
+                    // Check time property
+                    var timeProperty = timeInterfaceType.GetProperty("time");
+                    if (timeProperty != null)
+                    {
+                        Assert.IsNotNull(timeProperty);
+                        Assert.AreEqual(typeof(float), timeProperty.PropertyType);
+                        Assert.IsTrue(timeProperty.CanRead);
+                        Assert.IsFalse(timeProperty.CanWrite);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_TimestampConversion_ShouldWork()
+        {
+            // Test timestamp conversion logic concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test time conversion patterns used in HeaderSerializer
+                float testTime = 1234.567890f;
+                
+                // Test seconds extraction (truncation)
+                var seconds = (int)System.Math.Truncate(testTime);
+                Assert.AreEqual(1234, seconds);
+                
+                // Test nanoseconds calculation
+                var nanoseconds = (uint)((testTime - seconds) * 1e+9);
+                Assert.Greater(nanoseconds, 0u);
+                Assert.Less(nanoseconds, 1000000000u); // Should be less than 1 second in nanoseconds
+                
+                // Test precision
+                var reconstructedTime = seconds + (nanoseconds / 1e+9);
+                Assert.AreEqual(testTime, reconstructedTime, 1e-6f);
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_FrameIdHandling_ShouldWork()
+        {
+            // Test frame ID handling concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var frameIds = new string[] { 
+                    "base_link", "odom", "map", "laser", "camera_link", "" 
+                };
+                
+                foreach (var frameId in frameIds)
+                {
+                    Assert.IsNotNull(frameId); // Should not be null
+                    Assert.IsTrue(frameId.Length >= 0); // Should have valid length
+                    
+                    // Test frame ID validation patterns
+                    if (!string.IsNullOrEmpty(frameId))
+                    {
+                        Assert.IsFalse(frameId.Contains(" ")); // Should not contain spaces for ROS compatibility
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void HeaderSerializer_ROSMessageCreation_ShouldWork()
+        {
+            // Test ROS message creation concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test HeaderMsg creation pattern
+                var headerCreated = false;
+                var frameIdSet = false;
+                var stampSet = false;
+                
+                // Simulate HeaderMsg creation
+                if (!headerCreated) { headerCreated = true; }
+                if (!frameIdSet) { frameIdSet = true; }
+                if (!stampSet) { stampSet = true; }
+                
+                Assert.IsTrue(headerCreated);
+                Assert.IsTrue(frameIdSet);
+                Assert.IsTrue(stampSet);
+            });
+        }
+    }
+}
