+++ Assets/UnitySensors/Tests/Editor/HeaderSerializerTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8d6c9ba04bad9b139bcc35fdd4f855a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
