+++ Assets/UnitySensors/Tests/Editor/IPointInterfaceTests.cs
@@ -1,0 +1,195 @@
+using NUnit.Framework;
+using Unity.Mathematics;
+
+namespace UnitySensors.Tests.Editor
+{
+    [TestFixture]
+    public class IPointInterfaceTests
+    {
+        [Test]
+        public void IPointInterface_ReflectionAccess_ShouldBeAccessible()
+        {
+            // Test that IPointInterface can be accessed via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.Interface.Sensor.PointCloud.IPointInterface, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    Assert.IsTrue(type.IsInterface);
+                    Assert.IsTrue(type.IsPublic);
+                }
+            });
+        }
+
+        [Test]
+        public void IPointInterface_PositionProperty_ShouldBeAccessible()
+        {
+            // Test that position property is properly defined
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.Interface.Sensor.PointCloud.IPointInterface, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    var positionProperty = type.GetProperty("position");
+                    if (positionProperty != null)
+                    {
+                        Assert.IsNotNull(positionProperty);
+                        Assert.AreEqual(typeof(float3), positionProperty.PropertyType);
+                        Assert.IsTrue(positionProperty.CanRead);
+                        Assert.IsTrue(positionProperty.CanWrite);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void IPointInterface_ImplementingClasses_ShouldBeAccessible()
+        {
+            // Test that implementing classes can be found
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var pointTypes = new[] {
+                    "UnitySensors.DataType.PointCloud.PointXYZ, UnitySensorsRuntime",
+                    "UnitySensors.DataType.PointCloud.PointXYZI, UnitySensorsRuntime",
+                    "UnitySensors.DataType.PointCloud.PointXYZRGB, UnitySensorsRuntime"
+                };
+                
+                foreach (var typeName in pointTypes)
+                {
+                    var type = System.Type.GetType(typeName);
+                    if (type != null)
+                    {
+                        Assert.IsTrue(type.IsValueType); // Should be structs
+                        
+                        // Check if implements IPointInterface
+                        var interfaces = type.GetInterfaces();
+                        bool implementsIPointInterface = false;
+                        
+                        foreach (var iface in interfaces)
+                        {
+                            if (iface.Name.Contains("IPointInterface"))
+                            {
+                                implementsIPointInterface = true;
+                                break;
+                            }
+                        }
+                        
+                        if (interfaces.Length > 0) Assert.IsTrue(implementsIPointInterface);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void IPointInterface_Float3Operations_ShouldWork()
+        {
+            // Test float3 operations for position property
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test float3 creation and manipulation
+                var position1 = new float3(1.0f, 2.0f, 3.0f);
+                var position2 = new float3(4.0f, 5.0f, 6.0f);
+                
+                // Test basic operations
+                var sum = position1 + position2;
+                var difference = position2 - position1;
+                var scaled = position1 * 2.0f;
+                
+                Assert.AreEqual(new float3(5.0f, 7.0f, 9.0f), sum);
+                Assert.AreEqual(new float3(3.0f, 3.0f, 3.0f), difference);
+                Assert.AreEqual(new float3(2.0f, 4.0f, 6.0f), scaled);
+                
+                // Test magnitude
+                var magnitude = math.length(position1);
+                var expectedMagnitude = math.sqrt(1.0f + 4.0f + 9.0f);
+                Assert.AreEqual(expectedMagnitude, magnitude, 0.001f);
+            });
+        }
+
+        [Test]
+        public void IPointInterface_PositionSetter_ShouldWork()
+        {
+            // Test position property setter concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test position assignment patterns
+                var originalPosition = new float3(0.0f, 0.0f, 0.0f);
+                var newPosition = new float3(1.0f, 2.0f, 3.0f);
+                
+                // Simulate position assignment
+                var currentPosition = originalPosition;
+                currentPosition = newPosition;
+                
+                Assert.AreEqual(newPosition, currentPosition);
+                Assert.AreNotEqual(originalPosition, currentPosition);
+                
+                // Test component-wise assignment
+                currentPosition.x = 10.0f;
+                Assert.AreEqual(10.0f, currentPosition.x);
+                Assert.AreEqual(2.0f, currentPosition.y);
+                Assert.AreEqual(3.0f, currentPosition.z);
+            });
+        }
+
+        [Test]
+        public void IPointInterface_UnityMathematicsCompatibility_ShouldWork()
+        {
+            // Test Unity.Mathematics compatibility
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test common Unity.Mathematics operations
+                var position = new float3(1.0f, 0.0f, 0.0f);
+                
+                // Test normalization
+                var normalized = math.normalize(position);
+                Assert.AreEqual(1.0f, math.length(normalized), 0.001f);
+                
+                // Test distance calculation
+                var otherPosition = new float3(0.0f, 1.0f, 0.0f);
+                var distance = math.distance(position, otherPosition);
+                Assert.AreEqual(math.sqrt(2.0f), distance, 0.001f);
+                
+                // Test dot product
+                var dot = math.dot(position, otherPosition);
+                Assert.AreEqual(0.0f, dot, 0.001f); // Perpendicular vectors
+                
+                // Test cross product
+                var cross = math.cross(position, otherPosition);
+                Assert.AreEqual(new float3(0.0f, 0.0f, 1.0f), cross);
+            });
+        }
+
+        [Test]
+        public void IPointInterface_MemoryLayout_ShouldBeEfficient()
+        {
+            // Test memory layout efficiency concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test that float3 is efficiently packed
+                var position = new float3(1.0f, 2.0f, 3.0f);
+                
+                // float3 should be 12 bytes (3 * 4 bytes)
+                // This is conceptual testing since we can't directly measure in reflection tests
+                Assert.AreEqual(1.0f, position.x);
+                Assert.AreEqual(2.0f, position.y);
+                Assert.AreEqual(3.0f, position.z);
+                
+                // Test that memory layout is sequential for SIMD operations
+                var positions = new float3[] {
+                    new float3(1.0f, 2.0f, 3.0f),
+                    new float3(4.0f, 5.0f, 6.0f),
+                    new float3(7.0f, 8.0f, 9.0f)
+                };
+                
+                Assert.AreEqual(3, positions.Length);
+                
+                // Each position should maintain component order
+                for (int i = 0; i < positions.Length; i++)
+                {
+                    Assert.IsTrue(positions[i].x < positions[i].y);
+                    Assert.IsTrue(positions[i].y < positions[i].z);
+                }
+            });
+        }
+    }
+}
