+++ Assets/UnitySensors/Tests/Editor/IPointInterfaceTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1cd20ede69c51db1cae60a809bd8c3a6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
