+++ Assets/UnitySensors/Tests/Editor/ITimeInterfaceTests.cs
@@ -1,0 +1,213 @@
+using NUnit.Framework;
+
+namespace UnitySensors.Tests.Editor
+{
+    [TestFixture]
+    public class ITimeInterfaceTests
+    {
+        [Test]
+        public void ITimeInterface_ReflectionAccess_ShouldBeAccessible()
+        {
+            // Test that ITimeInterface can be accessed via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.Interface.Std.ITimeInterface, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    Assert.IsTrue(type.IsInterface);
+                    Assert.IsTrue(type.IsPublic);
+                }
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_TimeProperty_ShouldBeAccessible()
+        {
+            // Test that time property is properly defined
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.Interface.Std.ITimeInterface, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    var timeProperty = type.GetProperty("time");
+                    if (timeProperty != null)
+                    {
+                        Assert.IsNotNull(timeProperty);
+                        Assert.AreEqual(typeof(float), timeProperty.PropertyType);
+                        Assert.IsTrue(timeProperty.CanRead);
+                        Assert.IsFalse(timeProperty.CanWrite); // Should be read-only
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_TimeValues_ShouldBeValid()
+        {
+            // Test time value validation concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var timeTests = new[] {
+                    new { time = 0.0f, isValid = true },        // Start time
+                    new { time = 1.0f, isValid = true },        // 1 second
+                    new { time = 60.0f, isValid = true },       // 1 minute
+                    new { time = 3600.0f, isValid = true },     // 1 hour
+                    new { time = -1.0f, isValid = false },      // Negative time (invalid)
+                    new { time = float.NaN, isValid = false },  // NaN (invalid)
+                    new { time = float.PositiveInfinity, isValid = false } // Infinity (invalid)
+                };
+                
+                foreach (var test in timeTests)
+                {
+                    bool actualValid = !float.IsNaN(test.time) && 
+                                     !float.IsInfinity(test.time) && 
+                                     test.time >= 0.0f;
+                    
+                    Assert.AreEqual(test.isValid, actualValid);
+                }
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_UnityTimeCompatibility_ShouldWork()
+        {
+            // Test Unity time compatibility concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test Unity Time properties (conceptual)
+                var unityTimeProperties = new[] {
+                    "Time.time", "Time.unscaledTime", "Time.fixedTime", 
+                    "Time.deltaTime", "Time.unscaledDeltaTime"
+                };
+                
+                foreach (var property in unityTimeProperties)
+                {
+                    Assert.IsNotNull(property);
+                    Assert.IsTrue(property.StartsWith("Time."));
+                }
+                
+                // Test time progression concepts
+                float startTime = 0.0f;
+                float deltaTime = 0.016f; // ~60 FPS
+                float currentTime = startTime;
+                
+                // Simulate time progression
+                for (int frame = 0; frame < 10; frame++)
+                {
+                    currentTime += deltaTime;
+                    Assert.Greater(currentTime, startTime);
+                    Assert.AreEqual(startTime + (frame + 1) * deltaTime, currentTime, 1e-6f);
+                }
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_TimeComparison_ShouldWork()
+        {
+            // Test time comparison operations
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                float time1 = 1.0f;
+                float time2 = 2.0f;
+                float time3 = 1.0f;
+                
+                // Test equality
+                Assert.AreEqual(time1, time3);
+                Assert.AreNotEqual(time1, time2);
+                
+                // Test ordering
+                Assert.Less(time1, time2);
+                Assert.Greater(time2, time1);
+                Assert.LessOrEqual(time1, time3);
+                Assert.GreaterOrEqual(time3, time1);
+                
+                // Test time difference
+                float timeDiff = time2 - time1;
+                Assert.AreEqual(1.0f, timeDiff);
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_TimeStampAccuracy_ShouldBeReasonable()
+        {
+            // Test timestamp accuracy concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test typical timestamp precision requirements
+                float baseTime = 1000.0f; // 1000 seconds
+                float highPrecisionDelta = 0.001f; // 1 millisecond
+                
+                float preciseTime = baseTime + highPrecisionDelta;
+                
+                // Should be able to distinguish millisecond differences
+                Assert.AreNotEqual(baseTime, preciseTime);
+                Assert.AreEqual(highPrecisionDelta, preciseTime - baseTime, 1e-4f);
+                
+                // Test microsecond precision
+                float microDelta = 0.000001f; // 1 microsecond
+                float microTime = baseTime + microDelta;
+                
+                // Float precision may not handle microseconds reliably at this scale
+                // Check if the addition had any effect, but allow for precision limitations
+                if (microTime != baseTime) {
+                    Assert.Greater(microTime, baseTime);
+                } else {
+                    // At this precision level, float may not distinguish the difference
+                    Assert.AreEqual(baseTime, microTime);
+                }
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_ROSTimeCompatibility_ShouldWork()
+        {
+            // Test ROS time compatibility concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test ROS time format conversion concepts
+                float unityTime = 1234.567890f;
+                
+                // Convert to ROS time format (seconds + nanoseconds)
+                int seconds = (int)System.Math.Truncate(unityTime);
+                uint nanoseconds = (uint)((unityTime - seconds) * 1e9);
+                
+                Assert.AreEqual(1234, seconds);
+                Assert.Greater(nanoseconds, 0u);
+                Assert.Less(nanoseconds, 1000000000u); // Less than 1 second in nanoseconds
+                
+                // Test reverse conversion
+                float reconstructedTime = seconds + (nanoseconds / 1e9f);
+                Assert.AreEqual(unityTime, reconstructedTime, 1e-6f);
+                
+                // Test ROS time concepts
+                var rosTimeFields = new[] { "sec", "nanosec" };
+                foreach (var field in rosTimeFields)
+                {
+                    Assert.IsNotNull(field);
+                    Assert.IsTrue(field.Length > 0);
+                }
+            });
+        }
+
+        [Test]
+        public void ITimeInterface_TimeDeltaCalculation_ShouldWork()
+        {
+            // Test time delta calculation concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                float previousTime = 10.0f;
+                float currentTime = 10.016f;
+                float timeDelta = currentTime - previousTime;
+                
+                Assert.AreEqual(0.016f, timeDelta, 1e-6f);
+                Assert.Greater(timeDelta, 0.0f);
+                
+                // Test time advancement
+                float nextTime = currentTime + timeDelta;
+                
+                Assert.Greater(nextTime, currentTime);
+                Assert.AreEqual(timeDelta, nextTime - currentTime, 1e-6f);
+            });
+        }
+    }
+}
