+++ Assets/UnitySensors/Tests/Editor/ITimeInterfaceTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a1b2e568ba2eaf716bd322c1e4a2df29
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
