+++ Assets/UnitySensors/Tests/Editor/LaserScanMsgSerializerTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 482d190dac50969ee89387fe02161e53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
