+++ Assets/UnitySensors/Tests/Editor/PointCloudBasicTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae09abb620f8c5ace8b66c4e75ef01ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
