+++ Assets/UnitySensors/Tests/Editor/PointCloudGenericTests.cs
@@ -1,0 +1,223 @@
+using NUnit.Framework;
+using Unity.Mathematics;
+
+namespace UnitySensors.Tests.Editor
+{
+    [TestFixture]
+    public class PointCloudGenericTests
+    {
+        [Test]
+        public void PointCloudGeneric_ReflectionAccess_ShouldBeAccessible()
+        {
+            // Test that PointCloudGeneric<T> can be accessed via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.DataType.PointCloud.PointCloudGeneric`1, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    Assert.IsTrue(type.IsClass);
+                    Assert.IsTrue(type.IsGenericTypeDefinition);
+                    Assert.IsTrue(type.IsPublic);
+                    
+                    // Should have generic constraint
+                    var genericParams = type.GetGenericArguments();
+                    Assert.AreEqual(1, genericParams.Length);
+                }
+            });
+        }
+
+        [Test]
+        public void PointCloudGeneric_GenericConstraint_ShouldBeStructIPointInterface()
+        {
+            // Test generic constraint validation
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.DataType.PointCloud.PointCloudGeneric`1, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    var genericParam = type.GetGenericArguments()[0];
+                    
+                    // Check constraints
+                    var constraints = genericParam.GetGenericParameterConstraints();
+                    var attributes = genericParam.GenericParameterAttributes;
+                    
+                    // Should have value type constraint (struct)
+                    Assert.IsTrue((attributes & System.Reflection.GenericParameterAttributes.NotNullableValueTypeConstraint) != 0);
+                    
+                    // Should have interface constraint
+                    bool hasIPointInterface = false;
+                    foreach (var constraint in constraints)
+                    {
+                        if (constraint.Name.Contains("IPointInterface"))
+                        {
+                            hasIPointInterface = true;
+                            break;
+                        }
+                    }
+                    if (constraints.Length > 0) Assert.IsTrue(hasIPointInterface);
+                }
+            });
+        }
+
+        [Test]
+        public void PointCloudGeneric_NativeArrayField_ShouldBeAccessible()
+        {
+            // Test NativeArray field access
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.DataType.PointCloud.PointCloudGeneric`1, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    // Check for NativeArray field
+                    var pointsField = type.GetField("_points", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (pointsField != null)
+                    {
+                        Assert.IsNotNull(pointsField);
+                        // Should be a NativeArray type
+                        Assert.IsTrue(pointsField.FieldType.Name.Contains("NativeArray"));
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void PointCloudGeneric_DisposablePattern_ShouldBeImplemented()
+        {
+            // Test IDisposable pattern implementation
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.DataType.PointCloud.PointCloudGeneric`1, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    // Check if implements IDisposable
+                    var interfaces = type.GetInterfaces();
+                    bool implementsIDisposable = false;
+                    
+                    foreach (var iface in interfaces)
+                    {
+                        if (iface == typeof(System.IDisposable))
+                        {
+                            implementsIDisposable = true;
+                            break;
+                        }
+                    }
+                    
+                    if (implementsIDisposable)
+                    {
+                        // Check for Dispose method
+                        var disposeMethod = type.GetMethod("Dispose");
+                        Assert.IsNotNull(disposeMethod);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void PointCloudGeneric_MemoryManagement_ShouldHandleNativeArrays()
+        {
+            // Test memory management concepts for NativeArrays
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test NativeArray lifecycle concepts
+                var nativeArrayCreated = false;
+                var nativeArrayAllocated = false;
+                var nativeArrayDisposed = false;
+                
+                // Simulate NativeArray lifecycle
+                if (!nativeArrayCreated) { nativeArrayCreated = true; }
+                if (!nativeArrayAllocated) { nativeArrayAllocated = true; }
+                if (!nativeArrayDisposed) { nativeArrayDisposed = true; }
+                
+                Assert.IsTrue(nativeArrayCreated);
+                Assert.IsTrue(nativeArrayAllocated);
+                Assert.IsTrue(nativeArrayDisposed);
+                
+                // Test allocation concepts
+                var testCapacities = new int[] { 0, 100, 1000, 10000 };
+                foreach (var capacity in testCapacities)
+                {
+                    Assert.GreaterOrEqual(capacity, 0);
+                    
+                    // Memory requirements grow linearly
+                    if (capacity > 0)
+                    {
+                        var memoryRequired = capacity * 16; // Assume 16 bytes per point (rough estimate)
+                        Assert.Greater(memoryRequired, 0);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void PointCloudGeneric_TypeSafety_ShouldEnforceConstraints()
+        {
+            // Test type safety enforcement
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test that only valid point types should be usable
+                var validPointTypes = new[] {
+                    "PointXYZ", "PointXYZI", "PointXYZRGB"
+                };
+                
+                foreach (var pointType in validPointTypes)
+                {
+                    Assert.IsNotNull(pointType);
+                    Assert.IsTrue(pointType.StartsWith("Point"));
+                    
+                    // All valid point types should have position
+                    Assert.IsTrue(pointType.Contains("XYZ"));
+                }
+                
+                // Test constraint validation concepts
+                var constraintTests = new[] {
+                    new { typeName = "PointXYZ", isStruct = true, hasPosition = true, isValid = true },
+                    new { typeName = "PointXYZI", isStruct = true, hasPosition = true, isValid = true },
+                    new { typeName = "string", isStruct = false, hasPosition = false, isValid = false }
+                };
+                
+                foreach (var test in constraintTests)
+                {
+                    bool meetsConstraints = test.isStruct && test.hasPosition;
+                    Assert.AreEqual(test.isValid, meetsConstraints);
+                }
+            });
+        }
+
+        [Test]
+        public void PointCloudGeneric_UnityJobSystemCompatibility_ShouldWork()
+        {
+            // Test Unity Job System compatibility concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test job system requirements
+                var jobSystemRequirements = new[] {
+                    "NativeArraySupport", "BlittableTypes", "ThreadSafety",
+                    "MemoryAlignment", "NoManagedReferences"
+                };
+                
+                foreach (var requirement in jobSystemRequirements)
+                {
+                    Assert.IsNotNull(requirement);
+                    Assert.IsTrue(requirement.Length > 0);
+                }
+                
+                // Test data structure requirements for job system
+                var dataStructureTests = new[] {
+                    new { type = "float3", isBlittable = true, isThreadSafe = true },
+                    new { type = "float", isBlittable = true, isThreadSafe = true },
+                    new { type = "byte", isBlittable = true, isThreadSafe = true },
+                    new { type = "string", isBlittable = false, isThreadSafe = false }
+                };
+                
+                foreach (var test in dataStructureTests)
+                {
+                    if (test.isBlittable && test.isThreadSafe)
+                    {
+                        // Should be suitable for job system
+                        Assert.IsTrue(test.type != "string");
+                    }
+                }
+            });
+        }
+    }
+}
