+++ Assets/UnitySensors/Tests/Editor/PointCloudGenericTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f381295601fa89b46b725dc9d709a61c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
