+++ Assets/UnitySensors/Tests/Editor/RosMsgPublisherTests.cs
@@ -1,0 +1,277 @@
+using NUnit.Framework;
+using UnityEngine;
+
+namespace UnitySensors.Tests.Editor
+{
+    [TestFixture]
+    public class RosMsgPublisherTests
+    {
+        [Test]
+        public void RosMsgPublisher_ReflectionAccess_ShouldBeAccessible()
+        {
+            // Test that RosMsgPublisher<T, TT> can be accessed via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    Assert.IsTrue(type.IsClass);
+                    Assert.IsFalse(type.IsAbstract);
+                    Assert.IsTrue(type.IsPublic);
+                    Assert.IsTrue(type.IsGenericTypeDefinition);
+                    
+                    // Check that it's a generic type with 2 parameters
+                    Assert.AreEqual(2, type.GetGenericArguments().Length);
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_Inheritance_ShouldExtendMonoBehaviour()
+        {
+            // Test that RosMsgPublisher properly inherits from MonoBehaviour
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    var baseType = type.BaseType;
+                    Assert.IsNotNull(baseType);
+                    Assert.AreEqual(typeof(MonoBehaviour), baseType);
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_GenericConstraints_ShouldBeCorrect()
+        {
+            // Test that generic constraints are properly defined
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    var genericArgs = type.GetGenericArguments();
+                    Assert.AreEqual(2, genericArgs.Length);
+                    
+                    // First parameter (T) should have constraints
+                    var firstParam = genericArgs[0];
+                    Assert.IsNotNull(firstParam);
+                    
+                    // Second parameter (TT) should have constraints
+                    var secondParam = genericArgs[1];
+                    Assert.IsNotNull(secondParam);
+                    
+                    // Check new() constraint on second parameter
+                    var attrs = secondParam.GenericParameterAttributes;
+                    Assert.IsTrue((attrs & System.Reflection.GenericParameterAttributes.DefaultConstructorConstraint) != 0);
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_ConfigurationFields_ShouldBeAccessible()
+        {
+            // Test that RosMsgPublisher configuration fields are properly defined
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    // Check _frequency field
+                    var frequencyField = type.GetField("_frequency", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (frequencyField != null)
+                    {
+                        Assert.IsNotNull(frequencyField);
+                        Assert.AreEqual(typeof(float), frequencyField.FieldType);
+                    }
+                    
+                    // Check _topicName field
+                    var topicNameField = type.GetField("_topicName", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (topicNameField != null)
+                    {
+                        Assert.IsNotNull(topicNameField);
+                        Assert.AreEqual(typeof(string), topicNameField.FieldType);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_MonoBehaviourMethods_ShouldBeAccessible()
+        {
+            // Test that MonoBehaviour lifecycle methods are properly implemented
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
+                if (type != null)
+                {
+                    // Check Start method
+                    var startMethod = type.GetMethod("Start", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (startMethod != null)
+                    {
+                        Assert.IsNotNull(startMethod);
+                        Assert.AreEqual(typeof(void), startMethod.ReturnType);
+                    }
+                    
+                    // Check Update method
+                    var updateMethod = type.GetMethod("Update", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (updateMethod != null)
+                    {
+                        Assert.IsNotNull(updateMethod);
+                        Assert.AreEqual(typeof(void), updateMethod.ReturnType);
+                    }
+                    
+                    // Check OnDestroy method
+                    var onDestroyMethod = type.GetMethod("OnDestroy", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    if (onDestroyMethod != null)
+                    {
+                        Assert.IsNotNull(onDestroyMethod);
+                        Assert.AreEqual(typeof(void), onDestroyMethod.ReturnType);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_FrequencyCalculation_ShouldWork()
+        {
+            // Test frequency calculation concepts used in RosMsgPublisher
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                float frequency = 10.0f; // 10 Hz
+                float expectedInterval = 1.0f / frequency; // 0.1 seconds
+                
+                // Test interval calculation
+                float calculatedInterval = 1.0f / frequency;
+                Assert.AreEqual(expectedInterval, calculatedInterval, 1e-6f);
+                
+                // Test timing logic
+                float timeElapsed = 0.0f;
+                float deltaTime = 0.016f; // ~60 FPS
+                
+                // Simulate several frames
+                for (int i = 0; i < 10; i++)
+                {
+                    timeElapsed += deltaTime;
+                    
+                    if (timeElapsed >= expectedInterval)
+                    {
+                        // Should publish message
+                        Assert.IsTrue(timeElapsed >= expectedInterval);
+                        timeElapsed -= expectedInterval;
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_PublishingPattern_ShouldWork()
+        {
+            // Test publishing pattern concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test different frequencies
+                var frequencies = new float[] { 1.0f, 5.0f, 10.0f, 30.0f, 60.0f };
+                
+                foreach (var freq in frequencies)
+                {
+                    float interval = 1.0f / freq;
+                    Assert.IsTrue(interval > 0.0f);
+                    Assert.IsTrue(freq > 0.0f);
+                    
+                    // Test that higher frequencies have shorter intervals
+                    if (freq > 1.0f)
+                    {
+                        float baseInterval = 1.0f / 1.0f;
+                        Assert.IsTrue(interval < baseInterval);
+                    }
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_TopicNameValidation_ShouldWork()
+        {
+            // Test topic name validation concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var validTopicNames = new string[] { 
+                    "/scan", "/camera/image_raw", "/imu/data", "/odom", 
+                    "/tf", "/tf_static", "/sensor_msgs/laser_scan" 
+                };
+                
+                var invalidTopicNames = new string[] { 
+                    "", "   ", "invalid topic", "topic with spaces" 
+                };
+                
+                // Test valid topic names
+                foreach (var topicName in validTopicNames)
+                {
+                    Assert.IsNotNull(topicName);
+                    Assert.IsTrue(topicName.Length > 0);
+                    Assert.IsTrue(topicName.StartsWith("/") || topicName.Contains("/"));
+                }
+                
+                // Test invalid topic names
+                foreach (var topicName in invalidTopicNames)
+                {
+                    bool isValid = !string.IsNullOrWhiteSpace(topicName) && 
+                                  !topicName.Contains(" ") && 
+                                  topicName.Length > 0;
+                    Assert.IsFalse(isValid);
+                }
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_ROSConnectionPattern_ShouldWork()
+        {
+            // Test ROS connection pattern concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test ROS connection initialization pattern
+                var rosConnected = false;
+                var topicRegistered = false;
+                var publisherInitialized = false;
+                var serializerInitialized = false;
+                
+                // Simulate ROS connection lifecycle
+                if (!rosConnected) { rosConnected = true; }
+                if (!topicRegistered) { topicRegistered = true; }
+                if (!publisherInitialized) { publisherInitialized = true; }
+                if (!serializerInitialized) { serializerInitialized = true; }
+                
+                Assert.IsTrue(rosConnected);
+                Assert.IsTrue(topicRegistered);
+                Assert.IsTrue(publisherInitialized);
+                Assert.IsTrue(serializerInitialized);
+            });
+        }
+
+        [Test]
+        public void RosMsgPublisher_SerializationIntegration_ShouldWork()
+        {
+            // Test serialization integration concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test serialization workflow
+                var serializerCreated = false;
+                var serializerInitialized = false;
+                var messageSerialized = false;
+                var messagePublished = false;
+                
+                // Simulate serialization and publishing workflow
+                if (!serializerCreated) { serializerCreated = true; }
+                if (!serializerInitialized) { serializerInitialized = true; }
+                if (!messageSerialized) { messageSerialized = true; }
+                if (!messagePublished) { messagePublished = true; }
+                
+                Assert.IsTrue(serializerCreated);
+                Assert.IsTrue(serializerInitialized);
+                Assert.IsTrue(messageSerialized);
+                Assert.IsTrue(messagePublished);
+            });
+        }
+    }
+}
