+++ Assets/UnitySensors/Tests/Editor/RosMsgPublisherTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8b3cc43922f2aabc989d7f86f8110a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
