+++ Assets/UnitySensors/Tests/Editor/RosMsgSerializerTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a78444f4108cad2eeb11401804f60d39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
