+++ Assets/UnitySensors/Tests/Editor/ScanPatternBasicTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8d2eabefec7a41ac392d83b6b1a128b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
