+++ Assets/UnitySensors/Tests/Editor/ScanPatternGeneratorTests.cs
@@ -1,0 +1,272 @@
+using NUnit.Framework;
+using UnityEngine;
+using Unity.Mathematics;
+
+namespace UnitySensors.Tests.Editor
+{
+    [TestFixture]
+    public class ScanPatternGeneratorTests
+    {
+        [Test]
+        public void ScanPatternGenerator_ReflectionAccess_ShouldBeAccessible()
+        {
+            // Test that ScanPatternGenerator can be accessed via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.Utils.ScanPatternGenerator, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    Assert.IsTrue(type.IsClass);
+                    Assert.IsTrue(type.IsPublic);
+                }
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_EnumTypes_ShouldBeAccessible()
+        {
+            // Test that generator enum types are accessible
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test ScanPatternType enum
+                var scanPatternType = System.Type.GetType("UnitySensors.Utils.ScanPatternGenerator+ScanPatternType, UnitySensorsRuntime");
+                if (scanPatternType != null)
+                {
+                    Assert.IsTrue(scanPatternType.IsEnum);
+                    
+                    var enumValues = System.Enum.GetNames(scanPatternType);
+                    if (enumValues.Length > 0)
+                    {
+                        // Should have pattern types like Uniform, Custom, etc.
+                        Assert.Greater(enumValues.Length, 0);
+                    }
+                }
+                
+                // Test other generator-related enums
+                var generatorTypes = new[] {
+                    "ScanPatternType", "AngleUnit", "CoordinateSystem"
+                };
+                
+                foreach (var typeName in generatorTypes)
+                {
+                    Assert.IsNotNull(typeName);
+                    Assert.IsTrue(typeName.Length > 0);
+                }
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_CSVParsing_ShouldHandleBasicFormat()
+        {
+            // Test CSV parsing concepts for scan patterns
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test CSV format parsing concepts
+                var csvLines = new[] {
+                    "azimuth,elevation",
+                    "0.0,0.0",
+                    "1.0,0.0",
+                    "2.0,0.0",
+                    "-1.0,0.0"
+                };
+                
+                // Test header parsing
+                var header = csvLines[0];
+                var columns = header.Split(',');
+                Assert.AreEqual(2, columns.Length);
+                Assert.AreEqual("azimuth", columns[0]);
+                Assert.AreEqual("elevation", columns[1]);
+                
+                // Test data parsing
+                for (int i = 1; i < csvLines.Length; i++)
+                {
+                    var values = csvLines[i].Split(',');
+                    Assert.AreEqual(2, values.Length);
+                    
+                    // Should be parseable as floats
+                    Assert.IsTrue(float.TryParse(values[0], out float azimuth));
+                    Assert.IsTrue(float.TryParse(values[1], out float elevation));
+                    
+                    // Basic validation
+                    Assert.GreaterOrEqual(azimuth, -180.0f);
+                    Assert.LessOrEqual(azimuth, 180.0f);
+                    Assert.GreaterOrEqual(elevation, -90.0f);
+                    Assert.LessOrEqual(elevation, 90.0f);
+                }
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_AngleCalculations_ShouldBeCorrect()
+        {
+            // Test angle calculation concepts
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test uniform angle distribution
+                float minAngle = -45.0f;
+                float maxAngle = 45.0f;
+                int numPoints = 9; // Should give 10-degree increments
+                
+                float angleStep = (maxAngle - minAngle) / (numPoints - 1);
+                Assert.AreEqual(11.25f, angleStep, 0.01f);
+                
+                // Test angle sequence generation
+                var angles = new float[numPoints];
+                for (int i = 0; i < numPoints; i++)
+                {
+                    angles[i] = minAngle + i * angleStep;
+                }
+                
+                // Verify first and last angles
+                Assert.AreEqual(minAngle, angles[0], 0.001f);
+                Assert.AreEqual(maxAngle, angles[numPoints - 1], 0.001f);
+                
+                // Verify sequence is monotonic
+                for (int i = 1; i < numPoints; i++)
+                {
+                    Assert.Greater(angles[i], angles[i - 1]);
+                }
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_QuaternionRotations_ShouldBeCorrect()
+        {
+            // Test quaternion rotation calculations
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test rotation from angles
+                float azimuthDeg = 90.0f;
+                float elevationDeg = 45.0f;
+                
+                // Convert to radians
+                float azimuthRad = azimuthDeg * Mathf.PI / 180.0f;
+                float elevationRad = elevationDeg * Mathf.PI / 180.0f;
+                
+                // Create rotations
+                var azimuthRotation = Quaternion.AngleAxis(azimuthDeg, Vector3.up);
+                var elevationRotation = Quaternion.AngleAxis(elevationDeg, Vector3.right);
+                
+                // Test that rotations are valid (quaternions should be normalized)
+                Assert.AreEqual(1.0f, Mathf.Sqrt(azimuthRotation.x * azimuthRotation.x + azimuthRotation.y * azimuthRotation.y + azimuthRotation.z * azimuthRotation.z + azimuthRotation.w * azimuthRotation.w), 0.001f);
+                Assert.AreEqual(1.0f, Mathf.Sqrt(elevationRotation.x * elevationRotation.x + elevationRotation.y * elevationRotation.y + elevationRotation.z * elevationRotation.z + elevationRotation.w * elevationRotation.w), 0.001f);
+                
+                // Test rotation application
+                var forward = Vector3.forward;
+                var rotatedVector = azimuthRotation * forward;
+                
+                // 90-degree azimuth rotation should point to the right
+                Assert.AreEqual(Vector3.right.x, rotatedVector.normalized.x, 0.001f);
+                Assert.AreEqual(Vector3.right.y, rotatedVector.normalized.y, 0.001f);
+                Assert.AreEqual(Vector3.right.z, rotatedVector.normalized.z, 0.001f);
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_ArraySizeCalculation_ShouldBeCorrect()
+        {
+            // Test array size calculation for scan patterns
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test size calculation for different configurations
+                var sizeTests = new[] {
+                    new { azimuthPoints = 360, elevationPoints = 1, expectedSize = 360 },
+                    new { azimuthPoints = 180, elevationPoints = 90, expectedSize = 16200 },
+                    new { azimuthPoints = 8, elevationPoints = 4, expectedSize = 32 }
+                };
+                
+                foreach (var test in sizeTests)
+                {
+                    int calculatedSize = test.azimuthPoints * test.elevationPoints;
+                    Assert.AreEqual(test.expectedSize, calculatedSize);
+                    
+                    // Size should be positive
+                    Assert.Greater(calculatedSize, 0);
+                    
+                    // Size should match expected pattern
+                    Assert.AreEqual(test.azimuthPoints * test.elevationPoints, calculatedSize);
+                }
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_AngleRangeTracking_ShouldBeCorrect()
+        {
+            // Test angle range tracking during generation
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                // Test range tracking concepts
+                float minAzimuth = float.MaxValue;
+                float maxAzimuth = float.MinValue;
+                float minElevation = float.MaxValue;
+                float maxElevation = float.MinValue;
+                
+                // Sample angles
+                var testAngles = new[] {
+                    new { azimuth = -90.0f, elevation = -45.0f },
+                    new { azimuth = 0.0f, elevation = 0.0f },
+                    new { azimuth = 90.0f, elevation = 45.0f },
+                    new { azimuth = 45.0f, elevation = -30.0f }
+                };
+                
+                // Track ranges
+                foreach (var angle in testAngles)
+                {
+                    minAzimuth = Mathf.Min(minAzimuth, angle.azimuth);
+                    maxAzimuth = Mathf.Max(maxAzimuth, angle.azimuth);
+                    minElevation = Mathf.Min(minElevation, angle.elevation);
+                    maxElevation = Mathf.Max(maxElevation, angle.elevation);
+                }
+                
+                // Verify tracked ranges
+                Assert.AreEqual(-90.0f, minAzimuth);
+                Assert.AreEqual(90.0f, maxAzimuth);
+                Assert.AreEqual(-45.0f, minElevation);
+                Assert.AreEqual(45.0f, maxElevation);
+            });
+        }
+
+        [Test]
+        public void ScanPatternGenerator_SerializationFields_ShouldBeAccessible()
+        {
+            // Test serialization field access via reflection
+            // Act & Assert
+            Assert.DoesNotThrow(() => {
+                var type = System.Type.GetType("UnitySensors.Utils.ScanPatternGenerator, UnitySensorsRuntime");
+                if (type != null)
+                {
+                    // Check for configuration fields
+                    var fields = type.GetFields(System.Reflection.BindingFlags.Public | System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    
+                    bool hasConfigurationFields = fields.Length > 0;
+                    if (hasConfigurationFields)
+                    {
+                        // Should have fields for pattern generation
+                        Assert.Greater(fields.Length, 0);
+                        
+                        // Look for typical generator fields
+                        var expectedFieldTypes = new[] {
+                            typeof(float), typeof(int), typeof(string), typeof(bool)
+                        };
+                        
+                        bool hasExpectedTypes = false;
+                        foreach (var field in fields)
+                        {
+                            foreach (var expectedType in expectedFieldTypes)
+                            {
+                                if (field.FieldType == expectedType)
+                                {
+                                    hasExpectedTypes = true;
+                                    break;
+                                }
+                            }
+                            if (hasExpectedTypes) break;
+                        }
+                        
+                        Assert.IsTrue(hasExpectedTypes);
+                    }
+                }
+            });
+        }
+    }
+}
