+++ Assets/UnitySensors/Tests/Editor/ScanPatternGeneratorTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19b970258cc624d39bb854098a90c7bf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
