+++ Assets/UnitySensors/Tests/Editor/UnitySensorBasicTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54ed9c59814c3cbba9e81670756fb4c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
