+++ Assets/UnitySensors/Tests/Editor/UnitySensorsEditorTests.cs
@@ -1,2647 +1,68 @@
+// This file serves as the main test file for UnitySensors
+// All individual test classes have been moved to separate files:
+//
+// - BasicUnityTests.cs
+// - UnitySensorBasicTests.cs
+// - GeoCoordinateBasicTests.cs
+// - Vector3DBasicTests.cs
+// - ScanPatternBasicTests.cs
+// - PointCloudBasicTests.cs
+// - PointCloudGenericTests.cs
+// - IPointInterfaceTests.cs
+// - ITimeInterfaceTests.cs
+// - ScanPatternGeneratorTests.cs
+// - RosMsgSerializerTests.cs
+// - HeaderSerializerTests.cs
+// - LaserScanMsgSerializerTests.cs
+// - RosMsgPublisherTests.cs
+// - LaserScanMsgPublisherTests.cs
+//
+// This organization improves maintainability and makes the tests easier to review.
+// Each test class focuses on a specific component or functionality.
+
 using NUnit.Framework;
-using UnityEngine;
 
 namespace UnitySensors.Tests.Editor
 {
     [TestFixture]
-    public class BasicUnityTests
+    public class UnitySensorsEditorTestsInfo
     {
         [Test]
-        public void Vector3_Magnitude_ShouldCalculateCorrectly()
+        public void TestSuite_Info_ShouldProvideInformation()
         {
-            // Arrange
-            var vector = new Vector3(3.0f, 4.0f, 0.0f);
-            var expectedMagnitude = 5.0f;
-            
-            // Act
-            var magnitude = vector.magnitude;
-            
-            // Assert
-            Assert.AreEqual(expectedMagnitude, magnitude, 0.001f);
-        }
-
-        [Test]
-        public void Vector3_Normalize_ShouldReturnUnitVector()
-        {
-            // Arrange
-            var vector = new Vector3(3.0f, 4.0f, 0.0f);
-            
-            // Act
-            var normalized = vector.normalized;
-            
-            // Assert
-            Assert.AreEqual(1.0f, normalized.magnitude, 0.001f);
-        }
-
-        [Test]
-        public void Quaternion_Identity_ShouldBeNoRotation()
-        {
-            // Arrange & Act
-            var identity = Quaternion.identity;
-            
-            // Assert
-            Assert.AreEqual(0.0f, identity.x, 0.001f);
-            Assert.AreEqual(0.0f, identity.y, 0.001f);
-            Assert.AreEqual(0.0f, identity.z, 0.001f);
-            Assert.AreEqual(1.0f, identity.w, 0.001f);
-        }
-
-        [Test]
-        public void Color_Components_ShouldBeAccessible()
-        {
-            // Arrange
-            var color = new Color(0.8f, 0.6f, 0.4f, 1.0f);
-            
-            // Act & Assert
-            Assert.AreEqual(0.8f, color.r, 0.001f);
-            Assert.AreEqual(0.6f, color.g, 0.001f);
-            Assert.AreEqual(0.4f, color.b, 0.001f);
-            Assert.AreEqual(1.0f, color.a, 0.001f);
-        }
-
-        [Test]
-        public void Vector2Int_Constructor_ShouldSetCorrectValues()
-        {
-            // Arrange
-            var x = 640;
-            var y = 480;
-            
-            // Act
-            var vector = new Vector2Int(x, y);
-            
-            // Assert
-            Assert.AreEqual(x, vector.x);
-            Assert.AreEqual(y, vector.y);
-        }
-
-        [Test]
-        public void Transform_LocalPosition_ShouldDefaultToZero()
-        {
-            // Arrange
-            var gameObject = new GameObject("TestObject");
-            var transform = gameObject.transform;
-            
-            // Act
-            var position = transform.localPosition;
-            
-            // Assert
-            Assert.AreEqual(Vector3.zero, position);
-            
-            // Cleanup
-            Object.DestroyImmediate(gameObject);
-        }
-
-        [Test]
-        public void Transform_LocalRotation_ShouldDefaultToIdentity()
-        {
-            // Arrange
-            var gameObject = new GameObject("TestObject");
-            var transform = gameObject.transform;
-            
-            // Act
-            var rotation = transform.localRotation;
-            
-            // Assert
-            Assert.AreEqual(Quaternion.identity, rotation);
-            
-            // Cleanup
-            Object.DestroyImmediate(gameObject);
-        }
-
-        [Test]
-        public void GameObject_Name_ShouldBeSettable()
-        {
-            // Arrange
-            var expectedName = "TestSensorObject";
-            
-            // Act
-            var gameObject = new GameObject(expectedName);
-            
-            // Assert
-            Assert.AreEqual(expectedName, gameObject.name);
-            
-            // Cleanup
-            Object.DestroyImmediate(gameObject);
-        }
-
-        [Test]
-        public void Time_DeltaTime_ShouldBePositive()
-        {
-            // Act
-            var deltaTime = Time.deltaTime;
-            
-            // Assert
-            Assert.GreaterOrEqual(deltaTime, 0.0f);
-        }
-
-        [Test]
-        public void Mathf_Approximately_ShouldWorkCorrectly()
-        {
-            // Arrange
-            var a = 1.0f;
-            var b = 1.0001f;
-            
-            // Act & Assert
-            Assert.IsTrue(Mathf.Approximately(a, a));
-            Assert.IsFalse(Mathf.Approximately(a, b));
-        }
-    }
-
-    [TestFixture]
-    public class UnitySensorBasicTests
-    {
-        [Test]
-        public void UnitySensors_Assembly_ShouldBeAccessible()
-        {
-            // This test verifies that the UnitySensors assembly is properly referenced
+            // This test provides information about the test suite organization
             // Act & Assert
             Assert.DoesNotThrow(() => {
-                var assemblyName = typeof(UnityEngine.Vector3).Assembly.GetName().Name;
-                Assert.IsNotNull(assemblyName);
-            });
-        }
-
-        [Test]
-        public void Vector2Int_UnitySensorsCompatibility_ShouldWork()
-        {
-            // Test that basic Unity types work in UnitySensors context
-            // Arrange
-            var resolution = new Vector2Int(640, 480);
-            
-            // Act & Assert
-            Assert.AreEqual(640, resolution.x);
-            Assert.AreEqual(480, resolution.y);
-        }
-
-        [Test]
-        public void Float_HighPrecision_ShouldBeHandled()
-        {
-            // Test precision handling that might be used in sensors
-            // Arrange
-            var frequency = 30.0f;
-            var deltaTime = 1.0f / frequency;
-            
-            // Act
-            var calculatedFrequency = 1.0f / deltaTime;
-            
-            // Assert
-            Assert.AreEqual(frequency, calculatedFrequency, 0.001f);
-        }
-    }
-
-    [TestFixture]
-    public class GeoCoordinateBasicTests
-    {
-        [Test]
-        public void GeoCoordinate_Constructor_ShouldSetCorrectValues()
-        {
-            // Test the simplest possible usage of GeoCoordinate
-            // Act & Assert - just verify it doesn't throw
-            Assert.DoesNotThrow(() => {
-                // Try to instantiate if the class is accessible
-                var type = System.Type.GetType("UnitySensors.DataType.Geometry.GeoCoordinate, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var instance = System.Activator.CreateInstance(type, 35.0, 139.0, 0.0);
-                    Assert.IsNotNull(instance);
-                }
-            });
-        }
-
-        [Test]
-        public void GeoCoordinate_Reflection_ShouldBeAccessible()
-        {
-            // Test that the UnitySensors assembly can be accessed via reflection
-            // Act
-            var assemblies = System.AppDomain.CurrentDomain.GetAssemblies();
-            var unitySensorsAssembly = System.Array.Find(assemblies, a => a.GetName().Name.Contains("UnitySensors"));
-            
-            // Assert
-            if (unitySensorsAssembly != null)
-            {
-                Assert.IsNotNull(unitySensorsAssembly);
-                var types = unitySensorsAssembly.GetTypes();
-                Assert.Greater(types.Length, 0);
-            }
-            else
-            {
-                Assert.Pass("UnitySensors assembly not found - expected in current test setup");
-            }
-        }
-
-        [Test]
-        public void GeoCoordinate_ValidCoordinates_ShouldBeHandled()
-        {
-            // Test with known valid coordinate ranges
-            // Arrange
-            var validCoordinates = new[]
-            {
-                new { lat = 0.0, lon = 0.0, alt = 0.0 },           // Equator, Prime Meridian
-                new { lat = 90.0, lon = 0.0, alt = 0.0 },          // North Pole
-                new { lat = -90.0, lon = 0.0, alt = 0.0 },         // South Pole
-                new { lat = 35.6762, lon = 139.6503, alt = 10.0 }  // Tokyo
-            };
-
-            // Act & Assert
-            foreach (var coord in validCoordinates)
-            {
-                Assert.DoesNotThrow(() => {
-                    // Test that the coordinate values are within valid ranges
-                    Assert.That(coord.lat, Is.InRange(-90.0, 90.0));
-                    Assert.That(coord.lon, Is.InRange(-180.0, 180.0));
-                    Assert.That(coord.alt, Is.GreaterThanOrEqualTo(-11000.0)); // Mariana Trench depth
-                });
-            }
-        }
-    }
-
-    [TestFixture]
-    public class Vector3DBasicTests
-    {
-        [Test]
-        public void Vector3D_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that Vector3D can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Geometry.Vector3D, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    // Test default constructor
-                    var instance = System.Activator.CreateInstance(type);
-                    Assert.IsNotNull(instance);
-                    
-                    // Test parameterized constructor
-                    var paramInstance = System.Activator.CreateInstance(type, 1.0, 2.0, 3.0);
-                    Assert.IsNotNull(paramInstance);
-                }
-            });
-        }
-
-        [Test]
-        public void Vector3D_DoubleToFloatConversion_ShouldHandlePrecision()
-        {
-            // Test precision handling in Vector3D conversions
-            // Arrange
-            var highPrecisionValue = 123.456789012345;
-            var expectedFloatValue = (float)highPrecisionValue;
-            
-            // Act
-            var actualFloatValue = (float)highPrecisionValue;
-            
-            // Assert
-            Assert.AreEqual(expectedFloatValue, actualFloatValue, 1e-6f);
-        }
-
-        [Test]
-        public void Vector3D_UnityVector3Compatibility_ShouldWork()
-        {
-            // Test that Vector3D-style operations work with Unity Vector3
-            // Arrange
-            var unityVector = new Vector3(1.5f, 2.5f, 3.5f);
-            
-            // Act - simulate Vector3D operations
-            var doubleX = (double)unityVector.x;
-            var doubleY = (double)unityVector.y;
-            var doubleZ = (double)unityVector.z;
-            
-            // Assert
-            Assert.AreEqual(1.5, doubleX, 1e-6);
-            Assert.AreEqual(2.5, doubleY, 1e-6);
-            Assert.AreEqual(3.5, doubleZ, 1e-6);
-        }
-
-        [Test]
-        public void Vector3D_PrecisionComparison_ShouldShowDifferences()
-        {
-            // Test the precision differences between float and double
-            // Arrange
-            var highPrecisionValue = 123.456789012345;
-            var floatValue = (float)highPrecisionValue;
-            var backToDouble = (double)floatValue;
-            
-            // Act & Assert
-            Assert.That(highPrecisionValue, Is.Not.EqualTo(backToDouble).Within(1e-10));
-            Assert.AreEqual(highPrecisionValue, backToDouble, 1e-5); // Within float precision
-        }
-
-        [Test]
-        public void Vector3D_DefaultValues_ShouldBeZero()
-        {
-            // Test default initialization behavior
-            // Arrange
-            double defaultX = 0.0;
-            double defaultY = 0.0;
-            double defaultZ = 0.0;
-            
-            // Act & Assert
-            Assert.AreEqual(0.0, defaultX, 1e-15);
-            Assert.AreEqual(0.0, defaultY, 1e-15);
-            Assert.AreEqual(0.0, defaultZ, 1e-15);
-        }
-
-        [Test]
-        public void Vector3D_VectorOperations_ShouldWorkWithDoubles()
-        {
-            // Test vector-like operations with double precision
-            // Arrange
-            var x1 = 1.234567890123456;
-            var y1 = 2.345678901234567;
-            var z1 = 3.456789012345678;
-            
-            var x2 = 0.123456789012345;
-            var y2 = 0.234567890123456;
-            var z2 = 0.345678901234567;
-            
-            // Act - vector addition
-            var sumX = x1 + x2;
-            var sumY = y1 + y2;
-            var sumZ = z1 + z2;
-            
-            // Assert - Using practical tolerance for vector operations
-            Assert.AreEqual(1.358024679135801, sumX, 1e-4);
-            Assert.AreEqual(2.580246791358023, sumY, 1e-4);
-            Assert.AreEqual(3.802468913580245, sumZ, 1e-4);
-        }
-    }
-
-    [TestFixture]
-    public class ScanPatternBasicTests
-    {
-        [Test]
-        public void ScanPattern_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that ScanPattern can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.LiDAR.ScanPattern, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsNotNull(type);
-                    Assert.IsTrue(type.IsSubclassOf(typeof(ScriptableObject)));
-                }
-            });
-        }
-
-        [Test]
-        public void ScanPattern_AngleRange_ShouldBeValid()
-        {
-            // Test angle range validation logic
-            // Arrange
-            var minZenith = 0.0f;
-            var maxZenith = 180.0f;
-            var minAzimuth = -180.0f;
-            var maxAzimuth = 180.0f;
-            
-            // Act & Assert
-            Assert.That(minZenith, Is.InRange(0.0f, 180.0f));
-            Assert.That(maxZenith, Is.InRange(0.0f, 180.0f));
-            Assert.That(minAzimuth, Is.InRange(-180.0f, 180.0f));
-            Assert.That(maxAzimuth, Is.InRange(-180.0f, 180.0f));
-            Assert.LessOrEqual(minZenith, maxZenith);
-            Assert.LessOrEqual(minAzimuth, maxAzimuth);
-        }
-
-        [Test]
-        public void ScanPattern_Float3Array_ShouldBeInitializable()
-        {
-            // Test float3 array operations
-            // Arrange & Act
-            Assert.DoesNotThrow(() => {
-                var scans = new Unity.Mathematics.float3[4];
-                scans[0] = new Unity.Mathematics.float3(1.0f, 0.0f, 0.0f);
-                scans[1] = new Unity.Mathematics.float3(0.0f, 1.0f, 0.0f);
-                scans[2] = new Unity.Mathematics.float3(0.0f, 0.0f, 1.0f);
-                scans[3] = new Unity.Mathematics.float3(0.5f, 0.5f, 0.5f);
-                
-                // Assert
-                Assert.AreEqual(4, scans.Length);
-                Assert.AreEqual(1.0f, scans[0].x, 1e-6f);
-                Assert.AreEqual(1.0f, scans[1].y, 1e-6f);
-                Assert.AreEqual(1.0f, scans[2].z, 1e-6f);
-            });
-        }
-
-        [Test]
-        public void ScanPattern_AngleConversions_ShouldWorkCorrectly()
-        {
-            // Test angle conversion calculations
-            // Arrange
-            var degreesToRadians = Unity.Mathematics.math.PI / 180.0f;
-            var radiansToDegrees = 180.0f / Unity.Mathematics.math.PI;
-            
-            // Act
-            var angle90Deg = 90.0f;
-            var angle90Rad = angle90Deg * degreesToRadians;
-            var backToDegrees = angle90Rad * radiansToDegrees;
-            
-            // Assert
-            Assert.AreEqual(Unity.Mathematics.math.PI / 2.0f, angle90Rad, 1e-6f);
-            Assert.AreEqual(angle90Deg, backToDegrees, 1e-4f);
-        }
-
-        [Test]
-        public void ScanPattern_SphericalCoordinates_ShouldCalculateCorrectly()
-        {
-            // Test spherical coordinate calculations typical for LiDAR
-            // Arrange
-            var radius = 1.0f;
-            var zenithAngle = Unity.Mathematics.math.PI / 4.0f; // 45 degrees
-            var azimuthAngle = Unity.Mathematics.math.PI / 2.0f; // 90 degrees
-            
-            // Act - Convert spherical to cartesian
-            var x = radius * Unity.Mathematics.math.sin(zenithAngle) * Unity.Mathematics.math.cos(azimuthAngle);
-            var y = radius * Unity.Mathematics.math.sin(zenithAngle) * Unity.Mathematics.math.sin(azimuthAngle);
-            var z = radius * Unity.Mathematics.math.cos(zenithAngle);
-            
-            // Assert
-            Assert.AreEqual(0.0f, x, 1e-6f); // cos(90°) = 0
-            Assert.AreEqual(Unity.Mathematics.math.sqrt(2.0f) / 2.0f, y, 1e-6f); // sin(45°) * sin(90°)
-            Assert.AreEqual(Unity.Mathematics.math.sqrt(2.0f) / 2.0f, z, 1e-6f); // cos(45°)
-        }
-
-        [Test]
-        public void ScanPattern_ScriptableObjectFields_ShouldBeAccessible()
-        {
-            // Test that ScriptableObject fields are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.LiDAR.ScanPattern, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var scansField = type.GetField("scans");
-                    var sizeField = type.GetField("size");
-                    var minZenithField = type.GetField("minZenithAngle");
-                    var maxZenithField = type.GetField("maxZenithAngle");
-                    var minAzimuthField = type.GetField("minAzimuthAngle");
-                    var maxAzimuthField = type.GetField("maxAzimuthAngle");
-                    
-                    if (scansField != null) Assert.IsNotNull(scansField);
-                    if (sizeField != null) Assert.IsNotNull(sizeField);
-                    if (minZenithField != null) Assert.IsNotNull(minZenithField);
-                    if (maxZenithField != null) Assert.IsNotNull(maxZenithField);
-                    if (minAzimuthField != null) Assert.IsNotNull(minAzimuthField);
-                    if (maxAzimuthField != null) Assert.IsNotNull(maxAzimuthField);
-                }
-            });
-        }
-    }
-
-    [TestFixture]
-    public class PointCloudBasicTests
-    {
-        [Test]
-        public void PointXYZ_StructLayout_ShouldBeCorrect()
-        {
-            // Test PointXYZ structure layout and size
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud.PointXYZ, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsValueType); // Should be struct
-                    var interfaces = type.GetInterfaces();
-                    Assert.IsTrue(System.Array.Exists(interfaces, i => i.Name == "IPointInterface"));
-                    
-                    // Check expected size (3 floats = 12 bytes)
-                    var size = System.Runtime.InteropServices.Marshal.SizeOf(type);
-                    Assert.AreEqual(12, size);
-                }
-            });
-        }
-
-        [Test]
-        public void PointXYZI_StructLayout_ShouldBeCorrect()
-        {
-            // Test PointXYZI structure layout and size
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud.PointXYZI, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsValueType); // Should be struct
-                    var interfaces = type.GetInterfaces();
-                    Assert.IsTrue(System.Array.Exists(interfaces, i => i.Name == "IPointInterface"));
-                    
-                    // Check expected size (4 floats = 16 bytes)
-                    var size = System.Runtime.InteropServices.Marshal.SizeOf(type);
-                    Assert.AreEqual(16, size);
-                }
-            });
-        }
-
-        [Test]
-        public void PointXYZRGB_StructLayout_ShouldBeCorrect()
-        {
-            // Test PointXYZRGB structure layout and size
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud.PointXYZRGB, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsValueType); // Should be struct
-                    var interfaces = type.GetInterfaces();
-                    Assert.IsTrue(System.Array.Exists(interfaces, i => i.Name == "IPointInterface"));
-                    
-                    // Check expected size (3 floats + 4 bytes = 16 bytes)
-                    var size = System.Runtime.InteropServices.Marshal.SizeOf(type);
-                    Assert.AreEqual(16, size);
-                }
-            });
-        }
-
-        [Test]
-        public void PointTypes_Float3Position_ShouldBeInitializable()
-        {
-            // Test float3 position initialization for all point types
-            // Arrange
-            var testPosition = new Unity.Mathematics.float3(1.5f, 2.5f, 3.5f);
-            
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test basic float3 operations
-                Assert.AreEqual(1.5f, testPosition.x, 1e-6f);
-                Assert.AreEqual(2.5f, testPosition.y, 1e-6f);
-                Assert.AreEqual(3.5f, testPosition.z, 1e-6f);
-                
-                // Test magnitude calculation
-                var magnitude = Unity.Mathematics.math.length(testPosition);
-                var expectedMagnitude = Unity.Mathematics.math.sqrt(1.5f * 1.5f + 2.5f * 2.5f + 3.5f * 3.5f);
-                Assert.AreEqual(expectedMagnitude, magnitude, 1e-6f);
-            });
-        }
-
-        [Test]
-        public void PointXYZI_IntensityRange_ShouldBeValid()
-        {
-            // Test intensity value validation for LiDAR applications
-            // Arrange
-            var validIntensities = new float[] { 0.0f, 0.5f, 1.0f, 100.0f, 255.0f };
-            var invalidIntensities = new float[] { -1.0f, -100.0f };
-            
-            // Act & Assert
-            foreach (var intensity in validIntensities)
-            {
-                Assert.GreaterOrEqual(intensity, 0.0f, $"Intensity {intensity} should be non-negative");
-            }
-            
-            foreach (var intensity in invalidIntensities)
-            {
-                Assert.Less(intensity, 0.0f, $"Intensity {intensity} should be negative (invalid)");
-            }
-        }
-
-        [Test]
-        public void PointXYZRGB_ColorChannels_ShouldBeInByteRange()
-        {
-            // Test RGB color channel byte range validation
-            // Arrange
-            var validColors = new byte[] { 0, 128, 255 };
-            var testCombinations = new[]
-            {
-                new { r = (byte)255, g = (byte)0, b = (byte)0, a = (byte)255 },    // Red
-                new { r = (byte)0, g = (byte)255, b = (byte)0, a = (byte)255 },    // Green
-                new { r = (byte)0, g = (byte)0, b = (byte)255, a = (byte)255 },    // Blue
-                new { r = (byte)255, g = (byte)255, b = (byte)255, a = (byte)255 }, // White
-                new { r = (byte)0, g = (byte)0, b = (byte)0, a = (byte)0 }          // Black/Transparent
-            };
-            
-            // Act & Assert
-            foreach (var color in validColors)
-            {
-                Assert.That(color, Is.InRange(0, 255));
-            }
-            
-            foreach (var combo in testCombinations)
-            {
-                Assert.That(combo.r, Is.InRange(0, 255));
-                Assert.That(combo.g, Is.InRange(0, 255));
-                Assert.That(combo.b, Is.InRange(0, 255));
-                Assert.That(combo.a, Is.InRange(0, 255));
-            }
-        }
-
-        [Test]
-        public void PointTypes_MemoryEfficiency_ShouldBeOptimal()
-        {
-            // Test memory efficiency assumptions for point cloud processing
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test assumptions about point sizes for memory calculations
-                var pointXYZSize = 12;  // 3 floats
-                var pointXYZISize = 16; // 4 floats
-                var pointXYZRGBSize = 16; // 3 floats + 4 bytes (padded)
-                
-                // Calculate memory for typical point cloud sizes
-                var typicalPointCount = 65536; // 64K points
-                var xyzMemory = typicalPointCount * pointXYZSize;
-                var xyziMemory = typicalPointCount * pointXYZISize;
-                var xyzrgbMemory = typicalPointCount * pointXYZRGBSize;
-                
-                // Assert memory usage is reasonable for point clouds
-                Assert.Less(xyzMemory, 1024 * 1024); // < 1MB for 64K points
-                Assert.Less(xyziMemory, 1024 * 1024 * 2); // < 2MB for 64K points
-                Assert.Less(xyzrgbMemory, 1024 * 1024 * 2); // < 2MB for 64K points
-            });
-        }
-    }
-
-    [TestFixture]
-    public class PointCloudGenericTests
-    {
-        [Test]
-        public void PointCloudGeneric_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that PointCloud<T> generic class can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud`1, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsGenericTypeDefinition);
-                    Assert.IsTrue(type.IsClass);
-                    
-                    // Check that it implements IDisposable
-                    var interfaces = type.GetInterfaces();
-                    Assert.IsTrue(System.Array.Exists(interfaces, i => i == typeof(System.IDisposable)));
-                }
-            });
-        }
-
-        [Test]
-        public void PointCloudGeneric_GenericConstraint_ShouldBeStructIPointInterface()
-        {
-            // Test that the generic constraint is properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud`1, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var constraints = type.GetGenericArguments()[0].GetGenericParameterConstraints();
-                    var attrs = type.GetGenericArguments()[0].GenericParameterAttributes;
-                    
-                    // Should have struct constraint
-                    Assert.IsTrue((attrs & System.Reflection.GenericParameterAttributes.NotNullableValueTypeConstraint) != 0);
-                }
-            });
-        }
-
-        [Test]
-        public void PointCloudGeneric_NativeArrayField_ShouldBeAccessible()
-        {
-            // Test that the NativeArray<T> points field exists and is accessible
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud`1, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var pointsField = type.GetField("points");
-                    if (pointsField != null)
-                    {
-                        Assert.IsNotNull(pointsField);
-                        Assert.IsTrue(pointsField.FieldType.IsGenericType);
-                        
-                        // Check that it's a NativeArray type
-                        var fieldTypeName = pointsField.FieldType.GetGenericTypeDefinition().Name;
-                        Assert.IsTrue(fieldTypeName.Contains("NativeArray"));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void PointCloudGeneric_DisposablePattern_ShouldBeImplemented()
-        {
-            // Test that the IDisposable pattern is properly implemented
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.Sensor.PointCloud`1, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var disposeMethod = type.GetMethod("Dispose");
-                    if (disposeMethod != null)
-                    {
-                        Assert.IsNotNull(disposeMethod);
-                        Assert.AreEqual(typeof(void), disposeMethod.ReturnType);
-                        Assert.AreEqual(0, disposeMethod.GetParameters().Length);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void PointCloudGeneric_MemoryManagement_ShouldHandleNativeArrays()
-        {
-            // Test memory management concepts for NativeArray usage
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test NativeArray memory allocation patterns
-                var testSize = 1000;
-                var bytesPerPoint = 12; // PointXYZ size
-                var expectedMemory = testSize * bytesPerPoint;
-                
-                // Verify memory calculations are reasonable
-                Assert.Greater(testSize, 0);
-                Assert.Greater(bytesPerPoint, 0);
-                Assert.Greater(expectedMemory, 0);
-                Assert.Less(expectedMemory, 1024 * 1024); // Less than 1MB for 1K points
-            });
-        }
-
-        [Test]
-        public void PointCloudGeneric_TypeSafety_ShouldEnforceConstraints()
-        {
-            // Test type safety with different point types
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var validPointTypes = new[]
-                {
-                    "UnitySensors.DataType.Sensor.PointCloud.PointXYZ",
-                    "UnitySensors.DataType.Sensor.PointCloud.PointXYZI",
-                    "UnitySensors.DataType.Sensor.PointCloud.PointXYZRGB"
+                var testClasses = new[] {
+                    "BasicUnityTests - Basic Unity framework tests",
+                    "UnitySensorBasicTests - Basic UnitySensors functionality tests",
+                    "GeoCoordinateBasicTests - Geographical coordinate tests",
+                    "Vector3DBasicTests - 3D vector with double precision tests",
+                    "ScanPatternBasicTests - LiDAR scan pattern tests",
+                    "PointCloudBasicTests - Point cloud data structure tests",
+                    "PointCloudGenericTests - Generic point cloud wrapper tests",
+                    "IPointInterfaceTests - Point interface tests",
+                    "ITimeInterfaceTests - Time interface tests",
+                    "ScanPatternGeneratorTests - Scan pattern generator utility tests",
+                    "RosMsgSerializerTests - ROS message serializer base class tests",
+                    "HeaderSerializerTests - ROS header serializer tests",
+                    "LaserScanMsgSerializerTests - Laser scan message serializer tests",
+                    "RosMsgPublisherTests - ROS message publisher base class tests",
+                    "LaserScanMsgPublisherTests - Laser scan message publisher tests (FINAL TARGET)"
                 };
                 
-                foreach (var typeName in validPointTypes)
-                {
-                    var pointType = System.Type.GetType(typeName + ", UnitySensorsRuntime");
-                    if (pointType != null)
-                    {
-                        Assert.IsTrue(pointType.IsValueType); // Should be struct
-                        
-                        // Check if it implements IPointInterface
-                        var interfaces = pointType.GetInterfaces();
-                        Assert.IsTrue(System.Array.Exists(interfaces, i => i.Name == "IPointInterface"));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void PointCloudGeneric_UnityJobSystemCompatibility_ShouldWork()
-        {
-            // Test Unity Job System compatibility assumptions
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test that NativeArray concepts work as expected for Job System
-                // We can't actually create NativeArrays in editor tests, but we can test the concepts
+                Assert.AreEqual(15, testClasses.Length);
                 
-                // Test typical point cloud sizes for performance
-                var smallPointCloud = 1000;
-                var mediumPointCloud = 10000;
-                var largePointCloud = 100000;
-                
-                Assert.Greater(smallPointCloud, 0);
-                Assert.Greater(mediumPointCloud, smallPointCloud);
-                Assert.Greater(largePointCloud, mediumPointCloud);
-                
-                // Test that point sizes are reasonable for bulk operations
-                var pointXYZSize = 12;
-                var pointXYZISize = 16;
-                var pointXYZRGBSize = 16;
-                
-                Assert.AreEqual(12, pointXYZSize);
-                Assert.AreEqual(16, pointXYZISize);
-                Assert.AreEqual(16, pointXYZRGBSize);
-            });
-        }
-    }
-
-    [TestFixture]
-    public class ScanPatternGeneratorTests
-    {
-        [Test]
-        public void ScanPatternGenerator_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that ScanPatternGenerator can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.LiDAR.ScanPatternGenerator, UnitySensorsEditor");
-                if (type != null)
+                foreach (var testClass in testClasses)
                 {
-                    Assert.IsTrue(type.IsClass);
-                    Assert.IsTrue(type.IsSubclassOf(typeof(UnityEditor.EditorWindow)));
-                }
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_EnumTypes_ShouldBeAccessible()
-        {
-            // Test that nested enum types are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.LiDAR.ScanPatternGenerator, UnitySensorsEditor");
-                if (type != null)
-                {
-                    var nestedTypes = type.GetNestedTypes(System.Reflection.BindingFlags.NonPublic);
-                    
-                    // Check for Mode enum
-                    var modeType = System.Array.Find(nestedTypes, t => t.Name == "Mode");
-                    if (modeType != null)
-                    {
-                        Assert.IsTrue(modeType.IsEnum);
-                        var modeValues = System.Enum.GetNames(modeType);
-                        Assert.IsTrue(System.Array.Exists(modeValues, v => v == "FromCSV"));
-                        Assert.IsTrue(System.Array.Exists(modeValues, v => v == "FromSpecification"));
-                    }
-                    
-                    // Check for Direction enum
-                    var directionType = System.Array.Find(nestedTypes, t => t.Name == "Direction");
-                    if (directionType != null)
-                    {
-                        Assert.IsTrue(directionType.IsEnum);
-                        var directionValues = System.Enum.GetNames(directionType);
-                        Assert.IsTrue(System.Array.Exists(directionValues, v => v == "CW"));
-                        Assert.IsTrue(System.Array.Exists(directionValues, v => v == "CCW"));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_CSVParsing_ShouldHandleBasicFormat()
-        {
-            // Test CSV parsing logic with typical LiDAR data format
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical CSV header parsing
-                var testHeaders = new string[] 
-                {
-                    "azimuth,zenith,distance",
-                    "zenith_angle,azimuth_angle,range",
-                    "AZIMUTH,ZENITH,INTENSITY"
-                };
-                
-                foreach (var headerLine in testHeaders)
-                {
-                    var headers = headerLine.Split(',');
-                    
-                    int azimuth_index = -1;
-                    int zenith_index = -1;
-                    
-                    for (int c = 0; c < headers.Length; c++)
-                    {
-                        string header = headers[c].ToLower();
-                        if (header.Contains("zenith")) zenith_index = c;
-                        else if (header.Contains("azimuth")) azimuth_index = c;
-                    }
-                    
-                    Assert.GreaterOrEqual(azimuth_index, 0, $"Should find azimuth in: {headerLine}");
-                    Assert.GreaterOrEqual(zenith_index, 0, $"Should find zenith in: {headerLine}");
-                }
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_AngleCalculations_ShouldBeCorrect()
-        {
-            // Test angle calculation logic used in pattern generation
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test azimuth angle interpolation
-                var minAzimuth = -180.0f;
-                var maxAzimuth = 180.0f;
-                var resolution = 360;
-                
-                for (int i = 0; i < resolution; i++)
-                {
-                    // Test CW direction
-                    var azimuthCW = Mathf.Lerp(minAzimuth, maxAzimuth, (float)i / resolution);
-                    Assert.That(azimuthCW, Is.InRange(minAzimuth, maxAzimuth));
-                    
-                    // Test CCW direction  
-                    var azimuthCCW = Mathf.Lerp(minAzimuth, maxAzimuth, (float)(resolution - 1 - i) / resolution);
-                    Assert.That(azimuthCCW, Is.InRange(minAzimuth, maxAzimuth));
+                    Assert.IsNotNull(testClass);
+                    Assert.IsTrue(testClass.Length > 0);
+                    Assert.IsTrue(testClass.Contains("Tests"));
                 }
                 
-                // Test zenith angle offset
-                var zenithAngle = 15.0f;
-                var zenithOffset = 5.0f;
-                var adjustedZenith = zenithAngle - zenithOffset;
-                Assert.AreEqual(10.0f, adjustedZenith, 1e-6f);
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_QuaternionRotations_ShouldBeCorrect()
-        {
-            // Test quaternion rotation calculations for scan directions
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical LiDAR angles
-                var testAngles = new[]
-                {
-                    new { zenith = 0.0f, azimuth = 0.0f },      // Forward
-                    new { zenith = 0.0f, azimuth = 90.0f },     // Right
-                    new { zenith = 0.0f, azimuth = -90.0f },    // Left
-                    new { zenith = 15.0f, azimuth = 0.0f },     // Up 15°
-                    new { zenith = -15.0f, azimuth = 0.0f }     // Down 15°
-                };
-                
-                foreach (var angle in testAngles)
-                {
-                    // Test CSV mode rotation (positive zenith)
-                    var rotationCSV = Quaternion.Euler(angle.zenith, angle.azimuth, 0) * Vector3.forward;
-                    Assert.AreEqual(1.0f, rotationCSV.magnitude, 1e-6f, "Should be unit vector");
-                    
-                    // Test Specification mode rotation (negative zenith)
-                    var rotationSpec = Quaternion.Euler(-angle.zenith, angle.azimuth, 0) * Vector3.forward;
-                    Assert.AreEqual(1.0f, rotationSpec.magnitude, 1e-6f, "Should be unit vector");
-                }
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_ArraySizeCalculation_ShouldBeCorrect()
-        {
-            // Test array size calculation for scan pattern generation
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test specification mode size calculation
-                var zenithAngles = new float[] { -15.0f, 0.0f, 15.0f };
-                var azimuthResolution = 360;
-                var expectedSize = zenithAngles.Length * azimuthResolution;
-                
-                Assert.AreEqual(1080, expectedSize); // 3 * 360
-                
-                // Test CSV mode size calculation (lines - 2 for header and empty line)
-                var csvLines = new string[]
-                {
-                    "azimuth,zenith,distance",
-                    "0.0,0.0,10.0",
-                    "1.0,0.0,10.0",
-                    "2.0,0.0,10.0",
-                    "" // Empty line at end
-                };
-                var csvSize = csvLines.Length - 2;
-                Assert.AreEqual(3, csvSize);
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_AngleRangeTracking_ShouldBeCorrect()
-        {
-            // Test min/max angle range tracking during pattern generation
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var testAngles = new float[] { -15.0f, 0.0f, 15.0f, 30.0f };
-                
-                var minAngle = float.MaxValue;
-                var maxAngle = float.MinValue;
-                
-                foreach (var angle in testAngles)
-                {
-                    minAngle = Mathf.Min(minAngle, angle);
-                    maxAngle = Mathf.Max(maxAngle, angle);
-                }
-                
-                Assert.AreEqual(-15.0f, minAngle);
-                Assert.AreEqual(30.0f, maxAngle);
-            });
-        }
-
-        [Test]
-        public void ScanPatternGenerator_SerializationFields_ShouldBeAccessible()
-        {
-            // Test that serialized fields are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.DataType.LiDAR.ScanPatternGenerator, UnitySensorsEditor");
-                if (type != null)
-                {
-                    var expectedFields = new[]
-                    {
-                        "_direction",
-                        "_zenithAngles", 
-                        "_minAzimuthAngle",
-                        "_maxAzimuthAngle",
-                        "_azimuthAngleResolution",
-                        "_zenithAngleOffset"
-                    };
-                    
-                    foreach (var fieldName in expectedFields)
-                    {
-                        var field = type.GetField(fieldName, System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                        if (field != null)
-                        {
-                            Assert.IsNotNull(field);
-                            
-                            // Check for SerializeField attribute
-                            var attrs = field.GetCustomAttributes(typeof(SerializeField), false);
-                            Assert.Greater(attrs.Length, 0, $"Field {fieldName} should have SerializeField attribute");
-                        }
-                    }
-                }
-            });
-        }
-    }
-
-    [TestFixture]
-    public class IPointInterfaceTests
-    {
-        [Test]
-        public void IPointInterface_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that IPointInterface can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.Interface.Sensor.PointCloud.IPointInterface, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsInterface);
-                    Assert.IsTrue(type.IsPublic);
-                }
-            });
-        }
-
-        [Test]
-        public void IPointInterface_PositionProperty_ShouldBeAccessible()
-        {
-            // Test that the position property is properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.Interface.Sensor.PointCloud.IPointInterface, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var positionProperty = type.GetProperty("position");
-                    if (positionProperty != null)
-                    {
-                        Assert.IsNotNull(positionProperty);
-                        Assert.IsTrue(positionProperty.CanRead);
-                        Assert.IsTrue(positionProperty.CanWrite);
-                        
-                        // Check that it's a float3 type
-                        var propertyTypeName = positionProperty.PropertyType.Name;
-                        Assert.IsTrue(propertyTypeName.Contains("float3"));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void IPointInterface_Float3Operations_ShouldWork()
-        {
-            // Test float3 operations that would be used through the interface
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test float3 construction and operations
-                var position1 = new Unity.Mathematics.float3(1.0f, 2.0f, 3.0f);
-                var position2 = new Unity.Mathematics.float3(4.0f, 5.0f, 6.0f);
-                
-                // Test basic operations
-                Assert.AreEqual(1.0f, position1.x, 1e-6f);
-                Assert.AreEqual(2.0f, position1.y, 1e-6f);
-                Assert.AreEqual(3.0f, position1.z, 1e-6f);
-                
-                // Test vector operations
-                var sum = position1 + position2;
-                Assert.AreEqual(5.0f, sum.x, 1e-6f);
-                Assert.AreEqual(7.0f, sum.y, 1e-6f);
-                Assert.AreEqual(9.0f, sum.z, 1e-6f);
-                
-                // Test magnitude
-                var magnitude = Unity.Mathematics.math.length(position1);
-                var expectedMagnitude = Unity.Mathematics.math.sqrt(1.0f + 4.0f + 9.0f);
-                Assert.AreEqual(expectedMagnitude, magnitude, 1e-6f);
-            });
-        }
-
-        [Test]
-        public void IPointInterface_ImplementingClasses_ShouldBeAccessible()
-        {
-            // Test that known implementing classes can be found
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var pointTypes = new[]
-                {
-                    "UnitySensors.DataType.Sensor.PointCloud.PointXYZ",
-                    "UnitySensors.DataType.Sensor.PointCloud.PointXYZI",
-                    "UnitySensors.DataType.Sensor.PointCloud.PointXYZRGB"
-                };
-                
-                foreach (var typeName in pointTypes)
-                {
-                    var type = System.Type.GetType(typeName + ", UnitySensorsRuntime");
-                    if (type != null)
-                    {
-                        var interfaces = type.GetInterfaces();
-                        Assert.IsTrue(System.Array.Exists(interfaces, i => i.Name == "IPointInterface"));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void IPointInterface_PositionSetter_ShouldWork()
-        {
-            // Test position setter functionality through interface concept
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test position assignment patterns
-                var originalPosition = new Unity.Mathematics.float3(1.0f, 2.0f, 3.0f);
-                var newPosition = new Unity.Mathematics.float3(4.0f, 5.0f, 6.0f);
-                
-                // Test that positions can be compared
-                Assert.AreNotEqual(originalPosition.x, newPosition.x);
-                Assert.AreNotEqual(originalPosition.y, newPosition.y);
-                Assert.AreNotEqual(originalPosition.z, newPosition.z);
-                
-                // Test position copying
-                var copiedPosition = newPosition;
-                Assert.AreEqual(newPosition.x, copiedPosition.x, 1e-6f);
-                Assert.AreEqual(newPosition.y, copiedPosition.y, 1e-6f);
-                Assert.AreEqual(newPosition.z, copiedPosition.z, 1e-6f);
-            });
-        }
-
-        [Test]
-        public void IPointInterface_MemoryLayout_ShouldBeEfficient()
-        {
-            // Test memory layout assumptions for IPointInterface implementations
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test float3 memory size
-                var float3Size = System.Runtime.InteropServices.Marshal.SizeOf(typeof(Unity.Mathematics.float3));
-                Assert.AreEqual(12, float3Size); // 3 floats * 4 bytes each
-                
-                // Test typical point cloud sizes
-                var pointCount = 10000;
-                var positionMemory = pointCount * float3Size;
-                
-                // Should be reasonable for typical point clouds
-                Assert.Greater(positionMemory, 0);
-                Assert.Less(positionMemory, 1024 * 1024); // Less than 1MB for 10K points
-            });
-        }
-
-        [Test]
-        public void IPointInterface_UnityMathematicsCompatibility_ShouldWork()
-        {
-            // Test Unity.Mathematics compatibility for IPointInterface
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test Unity.Mathematics float3 operations
-                var position = new Unity.Mathematics.float3(1.0f, 2.0f, 3.0f);
-                
-                // Test normalization
-                var normalized = Unity.Mathematics.math.normalize(position);
-                var normalizedLength = Unity.Mathematics.math.length(normalized);
-                Assert.AreEqual(1.0f, normalizedLength, 1e-6f);
-                
-                // Test dot product
-                var otherPosition = new Unity.Mathematics.float3(1.0f, 0.0f, 0.0f);
-                var dot = Unity.Mathematics.math.dot(position, otherPosition);
-                Assert.AreEqual(1.0f, dot, 1e-6f);
-                
-                // Test cross product
-                var cross = Unity.Mathematics.math.cross(position, otherPosition);
-                Assert.AreEqual(0.0f, cross.x, 1e-6f);
-                Assert.AreEqual(3.0f, cross.y, 1e-6f);
-                Assert.AreEqual(-2.0f, cross.z, 1e-6f);
-            });
-        }
-    }
-
-    [TestFixture]
-    public class ITimeInterfaceTests
-    {
-        [Test]
-        public void ITimeInterface_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that ITimeInterface can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.Interface.Std.ITimeInterface, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsInterface);
-                    Assert.IsTrue(type.IsPublic);
-                }
-            });
-        }
-
-        [Test]
-        public void ITimeInterface_TimeProperty_ShouldBeAccessible()
-        {
-            // Test that the time property is properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.Interface.Std.ITimeInterface, UnitySensorsRuntime");
-                if (type != null)
-                {
-                    var timeProperty = type.GetProperty("time");
-                    if (timeProperty != null)
-                    {
-                        Assert.IsNotNull(timeProperty);
-                        Assert.IsTrue(timeProperty.CanRead);
-                        Assert.IsFalse(timeProperty.CanWrite); // Should be read-only
-                        
-                        // Check that it's a float type
-                        Assert.AreEqual(typeof(float), timeProperty.PropertyType);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void ITimeInterface_TimeValues_ShouldBeValid()
-        {
-            // Test time value validation concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical time values
-                var validTimes = new float[] { 0.0f, 1.0f, 10.0f, 100.0f, 1000.0f };
-                var invalidTimes = new float[] { -1.0f, -10.0f, float.NaN, float.PositiveInfinity };
-                
-                foreach (var time in validTimes)
-                {
-                    Assert.GreaterOrEqual(time, 0.0f, $"Time {time} should be non-negative");
-                    Assert.IsFalse(float.IsNaN(time), $"Time {time} should not be NaN");
-                    Assert.IsFalse(float.IsInfinity(time), $"Time {time} should not be infinite");
-                }
-                
-                foreach (var time in invalidTimes)
-                {
-                    Assert.IsTrue(time < 0.0f || float.IsNaN(time) || float.IsInfinity(time), 
-                        $"Time {time} should be invalid");
-                }
-            });
-        }
-
-        [Test]
-        public void ITimeInterface_UnityTimeCompatibility_ShouldWork()
-        {
-            // Test Unity time system compatibility
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test Unity time concepts
-                var unityTime = Time.time;
-                var unityFixedTime = Time.fixedTime;
-                var unityDeltaTime = Time.deltaTime;
-                
-                // Unity times should be valid
-                Assert.GreaterOrEqual(unityTime, 0.0f);
-                Assert.GreaterOrEqual(unityFixedTime, 0.0f);
-                Assert.GreaterOrEqual(unityDeltaTime, 0.0f);
-                
-                // Test time precision
-                Assert.IsTrue(unityTime is float);
-                Assert.IsTrue(unityFixedTime is float);
-                Assert.IsTrue(unityDeltaTime is float);
-            });
-        }
-
-        [Test]
-        public void ITimeInterface_TimeComparison_ShouldWork()
-        {
-            // Test time comparison operations
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var time1 = 10.0f;
-                var time2 = 20.0f;
-                var time3 = 10.0f;
-                
-                // Test basic comparisons
-                Assert.Less(time1, time2);
-                Assert.Greater(time2, time1);
-                Assert.AreEqual(time1, time3, 1e-6f);
-                
-                // Test time differences
-                var timeDiff = time2 - time1;
-                Assert.AreEqual(10.0f, timeDiff, 1e-6f);
-                
-                // Test time ordering
-                var times = new float[] { 5.0f, 1.0f, 10.0f, 3.0f };
-                System.Array.Sort(times);
-                Assert.AreEqual(1.0f, times[0], 1e-6f);
-                Assert.AreEqual(3.0f, times[1], 1e-6f);
-                Assert.AreEqual(5.0f, times[2], 1e-6f);
-                Assert.AreEqual(10.0f, times[3], 1e-6f);
-            });
-        }
-
-        [Test]
-        public void ITimeInterface_TimeStampAccuracy_ShouldBeReasonable()
-        {
-            // Test timestamp accuracy assumptions
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test float precision for typical time values
-                var startTime = 0.0f;
-                var frameTime = 1.0f / 60.0f; // 60 FPS
-                var endTime = startTime + frameTime;
-                
-                Assert.AreEqual(1.0f / 60.0f, frameTime, 1e-6f);
-                Assert.AreEqual(frameTime, endTime - startTime, 1e-6f);
-                
-                // Test accumulated time precision
-                var accumulatedTime = 0.0f;
-                for (int i = 0; i < 1000; i++)
-                {
-                    accumulatedTime += frameTime;
-                }
-                
-                var expectedTime = 1000.0f * frameTime;
-                Assert.AreEqual(expectedTime, accumulatedTime, 1e-3f); // Slightly larger tolerance for accumulation
-            });
-        }
-
-        [Test]
-        public void ITimeInterface_ROSTimeCompatibility_ShouldWork()
-        {
-            // Test ROS time concepts compatibility
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test ROS time epoch (seconds since Unix epoch)
-                var rosTime = 1609459200.0f; // 2021-01-01 00:00:00 UTC
-                var rosTimeNano = rosTime * 1e9f; // Convert to nanoseconds
-                
-                Assert.Greater(rosTime, 0.0f);
-                Assert.Greater(rosTimeNano, 0.0f);
-                
-                // Test time conversion concepts
-                var unityTime = Time.time;
-                var convertedTime = unityTime; // Direct assignment for interface
-                
-                Assert.AreEqual(unityTime, convertedTime, 1e-6f);
-                
-                // Test time synchronization concepts
-                var timeDelta = 0.016f; // Typical frame time
-                var nextTime = convertedTime + timeDelta;
-                
-                Assert.Greater(nextTime, convertedTime);
-                Assert.AreEqual(timeDelta, nextTime - convertedTime, 1e-6f);
-            });
-        }
-    }
-
-    [TestFixture]
-    public class RosMsgSerializerTests
-    {
-        [Test]
-        public void RosMsgSerializer_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that RosMsgSerializer<T> can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsGenericTypeDefinition);
-                    Assert.IsTrue(type.IsAbstract);
-                    Assert.IsTrue(type.IsClass);
-                    
-                    // Check System.Serializable attribute
-                    var attrs = type.GetCustomAttributes(typeof(System.SerializableAttribute), false);
-                    Assert.Greater(attrs.Length, 0, "Should have System.Serializable attribute");
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_GenericConstraint_ShouldBeMessage()
-        {
-            // Test that the generic constraint is properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var genericParam = type.GetGenericArguments()[0];
-                    var constraints = genericParam.GetGenericParameterConstraints();
-                    
-                    // Should have new() constraint
-                    var attrs = genericParam.GenericParameterAttributes;
-                    Assert.IsTrue((attrs & System.Reflection.GenericParameterAttributes.DefaultConstructorConstraint) != 0);
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_MessageProperty_ShouldBeAccessible()
-        {
-            // Test that the msg property is properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var msgProperty = type.GetProperty("msg");
-                    if (msgProperty != null)
-                    {
-                        Assert.IsNotNull(msgProperty);
-                        Assert.IsTrue(msgProperty.CanRead);
-                        Assert.IsFalse(msgProperty.CanWrite); // Should be read-only
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_AbstractMethods_ShouldBeAccessible()
-        {
-            // Test that abstract methods are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var serializeMethod = type.GetMethod("Serialize");
-                    if (serializeMethod != null)
-                    {
-                        Assert.IsNotNull(serializeMethod);
-                        Assert.IsTrue(serializeMethod.IsAbstract);
-                        Assert.IsTrue(serializeMethod.IsPublic);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_VirtualMethods_ShouldBeAccessible()
-        {
-            // Test that virtual methods are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var initMethod = type.GetMethod("Init");
-                    var onDestroyMethod = type.GetMethod("OnDestroy");
-                    
-                    if (initMethod != null)
-                    {
-                        Assert.IsNotNull(initMethod);
-                        Assert.IsTrue(initMethod.IsVirtual);
-                        Assert.IsTrue(initMethod.IsPublic);
-                        Assert.AreEqual(typeof(void), initMethod.ReturnType);
-                    }
-                    
-                    if (onDestroyMethod != null)
-                    {
-                        Assert.IsNotNull(onDestroyMethod);
-                        Assert.IsTrue(onDestroyMethod.IsVirtual);
-                        Assert.IsTrue(onDestroyMethod.IsPublic);
-                        Assert.AreEqual(typeof(void), onDestroyMethod.ReturnType);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_SerializationLifecycle_ShouldBeCorrect()
-        {
-            // Test the serialization lifecycle pattern
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical serialization lifecycle
-                var initCalled = false;
-                var serializeCalled = false;
-                var destroyCalled = false;
-                
-                // Simulate lifecycle
-                if (!initCalled) { initCalled = true; }
-                if (!serializeCalled) { serializeCalled = true; }
-                if (!destroyCalled) { destroyCalled = true; }
-                
-                // Verify lifecycle order
-                Assert.IsTrue(initCalled);
-                Assert.IsTrue(serializeCalled);
-                Assert.IsTrue(destroyCalled);
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_ROSMessageCompatibility_ShouldWork()
-        {
-            // Test ROS message compatibility concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test that ROS message patterns work
-                // We can't instantiate the abstract class, but we can test the concepts
-                
-                // Test message creation pattern
-                var messageCreated = false;
-                var messageInitialized = false;
-                var messageSerialized = false;
-                
-                // Simulate message lifecycle
-                if (!messageCreated) { messageCreated = true; }
-                if (!messageInitialized) { messageInitialized = true; }
-                if (!messageSerialized) { messageSerialized = true; }
-                
-                Assert.IsTrue(messageCreated);
-                Assert.IsTrue(messageInitialized);
-                Assert.IsTrue(messageSerialized);
-            });
-        }
-
-        [Test]
-        public void RosMsgSerializer_UnitySerializationCompatibility_ShouldWork()
-        {
-            // Test Unity serialization compatibility
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test Unity serialization concepts
-                var unitySerializable = true;
-                var systemSerializable = true;
-                
-                // Test that serialization attributes work
-                Assert.IsTrue(unitySerializable);
-                Assert.IsTrue(systemSerializable);
-                
-                // Test serialization in Unity context
-                var serializedInEditor = true;
-                var serializedInRuntime = true;
-                
-                Assert.IsTrue(serializedInEditor);
-                Assert.IsTrue(serializedInRuntime);
-            });
-        }
-    }
-
-    [TestFixture]
-    public class HeaderSerializerTests
-    {
-        [Test]
-        public void HeaderSerializer_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that HeaderSerializer can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsClass);
-                    Assert.IsFalse(type.IsAbstract);
-                    Assert.IsTrue(type.IsPublic);
-                    
-                    // Check System.Serializable attribute
-                    var attrs = type.GetCustomAttributes(typeof(System.SerializableAttribute), false);
-                    Assert.Greater(attrs.Length, 0, "Should have System.Serializable attribute");
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_Inheritance_ShouldExtendRosMsgSerializer()
-        {
-            // Test that HeaderSerializer properly inherits from RosMsgSerializer<HeaderMsg>
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var baseType = type.BaseType;
-                    Assert.IsNotNull(baseType);
-                    Assert.IsTrue(baseType.IsGenericType);
-                    
-                    // Check that it's RosMsgSerializer<HeaderMsg>
-                    var genericTypeDef = baseType.GetGenericTypeDefinition();
-                    var expectedBase = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                    if (expectedBase != null)
-                    {
-                        Assert.AreEqual(expectedBase, genericTypeDef);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_Fields_ShouldBeAccessible()
-        {
-            // Test that HeaderSerializer fields are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check _source field
-                    var sourceField = type.GetField("_source", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (sourceField != null)
-                    {
-                        Assert.IsNotNull(sourceField);
-                        Assert.AreEqual(typeof(UnityEngine.Object), sourceField.FieldType);
-                    }
-                    
-                    // Check _frame_id field
-                    var frameIdField = type.GetField("_frame_id", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (frameIdField != null)
-                    {
-                        Assert.IsNotNull(frameIdField);
-                        Assert.AreEqual(typeof(string), frameIdField.FieldType);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_InitMethod_ShouldBeOverridden()
-        {
-            // Test that Init method is properly overridden
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var initMethod = type.GetMethod("Init");
-                    if (initMethod != null)
-                    {
-                        Assert.IsNotNull(initMethod);
-                        Assert.IsTrue(initMethod.IsPublic);
-                        Assert.IsTrue(initMethod.IsVirtual);
-                        Assert.IsFalse(initMethod.IsAbstract);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_SerializeMethod_ShouldBeOverridden()
-        {
-            // Test that Serialize method is properly overridden
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var serializeMethod = type.GetMethod("Serialize");
-                    if (serializeMethod != null)
-                    {
-                        Assert.IsNotNull(serializeMethod);
-                        Assert.IsTrue(serializeMethod.IsPublic);
-                        Assert.IsTrue(serializeMethod.IsVirtual);
-                        Assert.IsFalse(serializeMethod.IsAbstract);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_ITimeInterface_ShouldBeAccessible()
-        {
-            // Test that ITimeInterface is properly used
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var timeInterfaceType = System.Type.GetType("UnitySensors.Interface.Std.ITimeInterface, UnitySensorsRuntime");
-                if (timeInterfaceType != null)
-                {
-                    Assert.IsTrue(timeInterfaceType.IsInterface);
-                    Assert.IsTrue(timeInterfaceType.IsPublic);
-                    
-                    // Check time property
-                    var timeProperty = timeInterfaceType.GetProperty("time");
-                    if (timeProperty != null)
-                    {
-                        Assert.IsNotNull(timeProperty);
-                        Assert.AreEqual(typeof(float), timeProperty.PropertyType);
-                        Assert.IsTrue(timeProperty.CanRead);
-                        Assert.IsFalse(timeProperty.CanWrite);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_TimestampConversion_ShouldWork()
-        {
-            // Test timestamp conversion logic concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test time conversion patterns used in HeaderSerializer
-                float testTime = 1234.567890f;
-                
-                // Test seconds extraction (truncation)
-                var seconds = (int)System.Math.Truncate(testTime);
-                Assert.AreEqual(1234, seconds);
-                
-                // Test nanoseconds calculation
-                var nanoseconds = (uint)((testTime - seconds) * 1e+9);
-                Assert.Greater(nanoseconds, 0u);
-                Assert.Less(nanoseconds, 1000000000u); // Should be less than 1 second in nanoseconds
-                
-                // Test precision
-                var reconstructedTime = seconds + (nanoseconds / 1e+9);
-                Assert.AreEqual(testTime, reconstructedTime, 1e-6f);
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_FrameIdHandling_ShouldWork()
-        {
-            // Test frame ID handling concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical frame ID patterns
-                var frameIds = new string[] { "base_link", "odom", "map", "laser", "camera_link", "" };
-                
-                foreach (var frameId in frameIds)
-                {
-                    Assert.IsNotNull(frameId); // Should not be null
-                    Assert.IsTrue(frameId.Length >= 0); // Should have valid length
-                    
-                    // Test frame ID validation patterns
-                    if (!string.IsNullOrEmpty(frameId))
-                    {
-                        Assert.IsFalse(frameId.Contains(" ")); // Should not contain spaces for ROS compatibility
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void HeaderSerializer_ROSMessageCreation_ShouldWork()
-        {
-            // Test ROS message creation concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test HeaderMsg creation pattern
-                var headerCreated = false;
-                var frameIdSet = false;
-                var stampSet = false;
-                
-                // Simulate HeaderMsg creation
-                if (!headerCreated) { headerCreated = true; }
-                if (!frameIdSet) { frameIdSet = true; }
-                if (!stampSet) { stampSet = true; }
-                
-                Assert.IsTrue(headerCreated);
-                Assert.IsTrue(frameIdSet);
-                Assert.IsTrue(stampSet);
-            });
-        }
-    }
-
-    [TestFixture]
-    public class LaserScanMsgSerializerTests
-    {
-        [Test]
-        public void LaserScanMsgSerializer_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that LaserScanMsgSerializer can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Sensor.LaserScanMsgSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsClass);
-                    Assert.IsFalse(type.IsAbstract);
-                    Assert.IsTrue(type.IsPublic);
-                    
-                    // Check System.Serializable attribute
-                    var attrs = type.GetCustomAttributes(typeof(System.SerializableAttribute), false);
-                    Assert.Greater(attrs.Length, 0, "Should have System.Serializable attribute");
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_Inheritance_ShouldExtendRosMsgSerializer()
-        {
-            // Test that LaserScanMsgSerializer properly inherits from RosMsgSerializer<LaserScanMsg>
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Sensor.LaserScanMsgSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var baseType = type.BaseType;
-                    Assert.IsNotNull(baseType);
-                    Assert.IsTrue(baseType.IsGenericType);
-                    
-                    // Check that it's RosMsgSerializer<LaserScanMsg>
-                    var genericTypeDef = baseType.GetGenericTypeDefinition();
-                    var expectedBase = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                    if (expectedBase != null)
-                    {
-                        Assert.AreEqual(expectedBase, genericTypeDef);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_ConfigurationFields_ShouldBeAccessible()
-        {
-            // Test that LaserScanMsgSerializer configuration fields are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Sensor.LaserScanMsgSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check _minRange field
-                    var minRangeField = type.GetField("_minRange", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (minRangeField != null)
-                    {
-                        Assert.IsNotNull(minRangeField);
-                        Assert.AreEqual(typeof(float), minRangeField.FieldType);
-                    }
-                    
-                    // Check _maxRange field
-                    var maxRangeField = type.GetField("_maxRange", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (maxRangeField != null)
-                    {
-                        Assert.IsNotNull(maxRangeField);
-                        Assert.AreEqual(typeof(float), maxRangeField.FieldType);
-                    }
-                    
-                    // Check _gaussianNoiseSigma field
-                    var noiseField = type.GetField("_gaussianNoiseSigma", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (noiseField != null)
-                    {
-                        Assert.IsNotNull(noiseField);
-                        Assert.AreEqual(typeof(float), noiseField.FieldType);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_DependencyFields_ShouldBeAccessible()
-        {
-            // Test that LaserScanMsgSerializer dependency fields are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Sensor.LaserScanMsgSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check _scanPattern field
-                    var scanPatternField = type.GetField("_scanPattern", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (scanPatternField != null)
-                    {
-                        Assert.IsNotNull(scanPatternField);
-                        // Should be ScanPattern type
-                        Assert.IsTrue(scanPatternField.FieldType.Name.Contains("ScanPattern"));
-                    }
-                    
-                    // Check _header field
-                    var headerField = type.GetField("_header", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (headerField != null)
-                    {
-                        Assert.IsNotNull(headerField);
-                        // Should be HeaderSerializer type
-                        Assert.IsTrue(headerField.FieldType.Name.Contains("HeaderSerializer"));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_Methods_ShouldBeOverridden()
-        {
-            // Test that LaserScanMsgSerializer methods are properly overridden
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Serializer.Sensor.LaserScanMsgSerializer, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check Init method
-                    var initMethod = type.GetMethod("Init");
-                    if (initMethod != null)
-                    {
-                        Assert.IsNotNull(initMethod);
-                        Assert.IsTrue(initMethod.IsPublic);
-                        Assert.IsTrue(initMethod.IsVirtual);
-                        Assert.IsFalse(initMethod.IsAbstract);
-                    }
-                    
-                    // Check Serialize method
-                    var serializeMethod = type.GetMethod("Serialize");
-                    if (serializeMethod != null)
-                    {
-                        Assert.IsNotNull(serializeMethod);
-                        Assert.IsTrue(serializeMethod.IsPublic);
-                        Assert.IsTrue(serializeMethod.IsVirtual);
-                        Assert.IsFalse(serializeMethod.IsAbstract);
-                    }
-                    
-                    // Check SetSource method
-                    var setSourceMethod = type.GetMethod("SetSource");
-                    if (setSourceMethod != null)
-                    {
-                        Assert.IsNotNull(setSourceMethod);
-                        Assert.IsTrue(setSourceMethod.IsPublic);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_RangeCalculation_ShouldWork()
-        {
-            // Test range calculation logic concepts used in LaserScanMsgSerializer
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test 2D range calculation from 3D points (x, y, z) -> sqrt(x² + z²)
-                var testPoints = new[] {
-                    new { x = 1.0f, y = 0.0f, z = 0.0f, expectedRange = 1.0f },
-                    new { x = 0.0f, y = 1.0f, z = 1.0f, expectedRange = 1.0f },
-                    new { x = 3.0f, y = 0.0f, z = 4.0f, expectedRange = 5.0f },
-                    new { x = 0.0f, y = 0.0f, z = 0.0f, expectedRange = 0.0f }
-                };
-                
-                foreach (var point in testPoints)
-                {
-                    var calculatedRange = System.Math.Sqrt(point.x * point.x + point.z * point.z);
-                    Assert.AreEqual(point.expectedRange, calculatedRange, 1e-6f);
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_RangeFiltering_ShouldWork()
-        {
-            // Test range filtering logic concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                float minRange = 0.5f;
-                float maxRange = 100.0f;
-                
-                var testRanges = new[] {
-                    new { range = 0.1f, shouldBeValid = false },  // Below minimum
-                    new { range = 0.5f, shouldBeValid = true },   // At minimum
-                    new { range = 50.0f, shouldBeValid = true },  // Within range
-                    new { range = 100.0f, shouldBeValid = true }, // At maximum
-                    new { range = 150.0f, shouldBeValid = false } // Above maximum
-                };
-                
-                foreach (var test in testRanges)
-                {
-                    bool isValid = test.range >= minRange && test.range <= maxRange;
-                    Assert.AreEqual(test.shouldBeValid, isValid);
-                    
-                    // Test NaN assignment for invalid ranges
-                    float filteredRange = isValid ? test.range : float.NaN;
-                    if (test.shouldBeValid)
-                    {
-                        Assert.AreEqual(test.range, filteredRange);
-                    }
-                    else
-                    {
-                        Assert.IsTrue(float.IsNaN(filteredRange));
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_GaussianNoise_ShouldWork()
-        {
-            // Test Gaussian noise application concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                float sigma = 0.1f;
-                float originalRange = 10.0f;
-                
-                // Test Box-Muller transform concept for Gaussian noise
-                var random = new System.Random(42); // Fixed seed for reproducibility
-                
-                // Generate multiple noise samples
-                var noiseSamples = new float[100];
-                for (int i = 0; i < noiseSamples.Length; i += 2)
-                {
-                    // Box-Muller transform
-                    double u1 = random.NextDouble();
-                    double u2 = random.NextDouble();
-                    double z0 = System.Math.Sqrt(-2.0 * System.Math.Log(u1)) * System.Math.Cos(2.0 * System.Math.PI * u2);
-                    
-                    noiseSamples[i] = (float)(z0 * sigma);
-                    if (i + 1 < noiseSamples.Length)
-                    {
-                        double z1 = System.Math.Sqrt(-2.0 * System.Math.Log(u1)) * System.Math.Sin(2.0 * System.Math.PI * u2);
-                        noiseSamples[i + 1] = (float)(z1 * sigma);
-                    }
-                }
-                
-                // Test that noise is applied correctly
-                for (int i = 0; i < noiseSamples.Length; i++)
-                {
-                    float noisyRange = originalRange + noiseSamples[i];
-                    Assert.IsTrue(noisyRange != originalRange || noiseSamples[i] == 0.0f);
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_LaserScanMsgStructure_ShouldWork()
-        {
-            // Test LaserScanMsg structure concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical LaserScanMsg structure
-                var laserScanCreated = false;
-                var headerSet = false;
-                var anglePropertiesSet = false;
-                var rangePropertiesSet = false;
-                var rangesArraySet = false;
-                var intensitiesArraySet = false;
-                
-                // Simulate LaserScanMsg creation and population
-                if (!laserScanCreated) { laserScanCreated = true; }
-                if (!headerSet) { headerSet = true; }
-                if (!anglePropertiesSet) { anglePropertiesSet = true; }
-                if (!rangePropertiesSet) { rangePropertiesSet = true; }
-                if (!rangesArraySet) { rangesArraySet = true; }
-                if (!intensitiesArraySet) { intensitiesArraySet = true; }
-                
-                Assert.IsTrue(laserScanCreated);
-                Assert.IsTrue(headerSet);
-                Assert.IsTrue(anglePropertiesSet);
-                Assert.IsTrue(rangePropertiesSet);
-                Assert.IsTrue(rangesArraySet);
-                Assert.IsTrue(intensitiesArraySet);
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgSerializer_IntensityMapping_ShouldWork()
-        {
-            // Test intensity mapping concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test intensity value handling
-                var testIntensities = new float[] { 0.0f, 0.5f, 1.0f, 100.0f, 255.0f };
-                
-                foreach (var intensity in testIntensities)
-                {
-                    // Test that intensity values are preserved
-                    Assert.AreEqual(intensity, intensity);
-                    Assert.IsTrue(intensity >= 0.0f);
-                    
-                    // Test intensity range validation
-                    var normalizedIntensity = intensity / 255.0f;
-                    Assert.IsTrue(normalizedIntensity >= 0.0f);
-                    Assert.IsTrue(normalizedIntensity <= (255.0f / 255.0f));
-                }
-            });
-        }
-    }
-
-    [TestFixture]
-    public class RosMsgPublisherTests
-    {
-        [Test]
-        public void RosMsgPublisher_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that RosMsgPublisher<T, TT> can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsClass);
-                    Assert.IsFalse(type.IsAbstract);
-                    Assert.IsTrue(type.IsPublic);
-                    Assert.IsTrue(type.IsGenericTypeDefinition);
-                    
-                    // Check that it's a generic type with 2 parameters
-                    Assert.AreEqual(2, type.GetGenericArguments().Length);
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_Inheritance_ShouldExtendMonoBehaviour()
-        {
-            // Test that RosMsgPublisher properly inherits from MonoBehaviour
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var baseType = type.BaseType;
-                    Assert.IsNotNull(baseType);
-                    Assert.AreEqual(typeof(MonoBehaviour), baseType);
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_GenericConstraints_ShouldBeCorrect()
-        {
-            // Test that generic constraints are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var genericArgs = type.GetGenericArguments();
-                    Assert.AreEqual(2, genericArgs.Length);
-                    
-                    // First parameter (T) should have constraints
-                    var firstParam = genericArgs[0];
-                    Assert.IsNotNull(firstParam);
-                    
-                    // Second parameter (TT) should have constraints
-                    var secondParam = genericArgs[1];
-                    Assert.IsNotNull(secondParam);
-                    
-                    // Check new() constraint on second parameter
-                    var attrs = secondParam.GenericParameterAttributes;
-                    Assert.IsTrue((attrs & System.Reflection.GenericParameterAttributes.DefaultConstructorConstraint) != 0);
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_ConfigurationFields_ShouldBeAccessible()
-        {
-            // Test that RosMsgPublisher configuration fields are properly defined
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check _frequency field
-                    var frequencyField = type.GetField("_frequency", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (frequencyField != null)
-                    {
-                        Assert.IsNotNull(frequencyField);
-                        Assert.AreEqual(typeof(float), frequencyField.FieldType);
-                    }
-                    
-                    // Check _topicName field
-                    var topicNameField = type.GetField("_topicName", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (topicNameField != null)
-                    {
-                        Assert.IsNotNull(topicNameField);
-                        Assert.AreEqual(typeof(string), topicNameField.FieldType);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_MonoBehaviourMethods_ShouldBeAccessible()
-        {
-            // Test that MonoBehaviour lifecycle methods are properly implemented
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check Start method
-                    var startMethod = type.GetMethod("Start", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (startMethod != null)
-                    {
-                        Assert.IsNotNull(startMethod);
-                        Assert.AreEqual(typeof(void), startMethod.ReturnType);
-                    }
-                    
-                    // Check Update method
-                    var updateMethod = type.GetMethod("Update", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (updateMethod != null)
-                    {
-                        Assert.IsNotNull(updateMethod);
-                        Assert.AreEqual(typeof(void), updateMethod.ReturnType);
-                    }
-                    
-                    // Check OnDestroy method
-                    var onDestroyMethod = type.GetMethod("OnDestroy", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                    if (onDestroyMethod != null)
-                    {
-                        Assert.IsNotNull(onDestroyMethod);
-                        Assert.AreEqual(typeof(void), onDestroyMethod.ReturnType);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_FrequencyCalculation_ShouldWork()
-        {
-            // Test frequency calculation concepts used in RosMsgPublisher
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                float frequency = 10.0f; // 10 Hz
-                float expectedInterval = 1.0f / frequency; // 0.1 seconds
-                
-                // Test interval calculation
-                float calculatedInterval = 1.0f / frequency;
-                Assert.AreEqual(expectedInterval, calculatedInterval, 1e-6f);
-                
-                // Test timing logic
-                float timeElapsed = 0.0f;
-                float deltaTime = 0.016f; // ~60 FPS
-                
-                // Simulate several frames
-                for (int i = 0; i < 10; i++)
-                {
-                    timeElapsed += deltaTime;
-                    
-                    if (timeElapsed >= expectedInterval)
-                    {
-                        // Should publish message
-                        Assert.IsTrue(timeElapsed >= expectedInterval);
-                        timeElapsed -= expectedInterval;
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_PublishingPattern_ShouldWork()
-        {
-            // Test publishing pattern concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test different frequencies
-                var frequencies = new float[] { 1.0f, 5.0f, 10.0f, 30.0f, 60.0f };
-                
-                foreach (var freq in frequencies)
-                {
-                    float interval = 1.0f / freq;
-                    Assert.IsTrue(interval > 0.0f);
-                    Assert.IsTrue(freq > 0.0f);
-                    
-                    // Test that higher frequencies have shorter intervals
-                    if (freq > 1.0f)
-                    {
-                        float baseInterval = 1.0f / 1.0f;
-                        Assert.IsTrue(interval < baseInterval);
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_TopicNameValidation_ShouldWork()
-        {
-            // Test topic name validation concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var validTopicNames = new string[] { 
-                    "/scan", "/camera/image_raw", "/imu/data", "/odom", 
-                    "/tf", "/tf_static", "/sensor_msgs/laser_scan" 
-                };
-                
-                var invalidTopicNames = new string[] { 
-                    "", "   ", "invalid topic", "topic with spaces" 
-                };
-                
-                // Test valid topic names
-                foreach (var topicName in validTopicNames)
-                {
-                    Assert.IsNotNull(topicName);
-                    Assert.IsTrue(topicName.Length > 0);
-                    Assert.IsTrue(topicName.StartsWith("/") || topicName.Contains("/"));
-                }
-                
-                // Test invalid topic names
-                foreach (var topicName in invalidTopicNames)
-                {
-                    bool isValid = !string.IsNullOrWhiteSpace(topicName) && 
-                                  !topicName.Contains(" ") && 
-                                  topicName.Length > 0;
-                    Assert.IsFalse(isValid);
-                }
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_ROSConnectionPattern_ShouldWork()
-        {
-            // Test ROS connection pattern concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test ROS connection initialization pattern
-                var rosConnected = false;
-                var topicRegistered = false;
-                var publisherInitialized = false;
-                var serializerInitialized = false;
-                
-                // Simulate ROS connection lifecycle
-                if (!rosConnected) { rosConnected = true; }
-                if (!topicRegistered) { topicRegistered = true; }
-                if (!publisherInitialized) { publisherInitialized = true; }
-                if (!serializerInitialized) { serializerInitialized = true; }
-                
-                Assert.IsTrue(rosConnected);
-                Assert.IsTrue(topicRegistered);
-                Assert.IsTrue(publisherInitialized);
-                Assert.IsTrue(serializerInitialized);
-            });
-        }
-
-        [Test]
-        public void RosMsgPublisher_SerializationIntegration_ShouldWork()
-        {
-            // Test serialization integration concepts
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test serialization workflow
-                var serializerCreated = false;
-                var serializerInitialized = false;
-                var messageSerialized = false;
-                var messagePublished = false;
-                
-                // Simulate serialization and publishing workflow
-                if (!serializerCreated) { serializerCreated = true; }
-                if (!serializerInitialized) { serializerInitialized = true; }
-                if (!messageSerialized) { messageSerialized = true; }
-                if (!messagePublished) { messagePublished = true; }
-                
-                Assert.IsTrue(serializerCreated);
-                Assert.IsTrue(serializerInitialized);
-                Assert.IsTrue(messageSerialized);
-                Assert.IsTrue(messagePublished);
-            });
-        }
-    }
-
-    [TestFixture]
-    public class LaserScanMsgPublisherTests
-    {
-        [Test]
-        public void LaserScanMsgPublisher_ReflectionAccess_ShouldBeAccessible()
-        {
-            // Test that LaserScanMsgPublisher can be accessed via reflection
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.Sensor.LaserScanMsgPublisher, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    Assert.IsTrue(type.IsClass);
-                    Assert.IsFalse(type.IsAbstract);
-                    Assert.IsTrue(type.IsPublic);
-                    Assert.IsFalse(type.IsGenericTypeDefinition);
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_Inheritance_ShouldExtendRosMsgPublisher()
-        {
-            // Test that LaserScanMsgPublisher properly inherits from RosMsgPublisher<LaserScanMsgSerializer, LaserScanMsg>
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.Sensor.LaserScanMsgPublisher, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    var baseType = type.BaseType;
-                    Assert.IsNotNull(baseType);
-                    Assert.IsTrue(baseType.IsGenericType);
-                    
-                    // Check that it's RosMsgPublisher<LaserScanMsgSerializer, LaserScanMsg>
-                    var genericTypeDef = baseType.GetGenericTypeDefinition();
-                    var expectedBase = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                    if (expectedBase != null)
-                    {
-                        Assert.AreEqual(expectedBase, genericTypeDef);
-                    }
-                    
-                    // Check generic type arguments
-                    var genericArgs = baseType.GetGenericArguments();
-                    Assert.AreEqual(2, genericArgs.Length);
-                    
-                    // First argument should be LaserScanMsgSerializer
-                    Assert.IsTrue(genericArgs[0].Name.Contains("LaserScanMsgSerializer"));
-                    
-                    // Second argument should be LaserScanMsg
-                    Assert.IsTrue(genericArgs[1].Name.Contains("LaserScanMsg"));
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_MonoBehaviourIntegration_ShouldWork()
-        {
-            // Test that LaserScanMsgPublisher integrates with Unity's MonoBehaviour system
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                var type = System.Type.GetType("UnitySensors.ROS.Publisher.Sensor.LaserScanMsgPublisher, UnitySensorsROSRuntime");
-                if (type != null)
-                {
-                    // Check that it ultimately derives from MonoBehaviour
-                    var currentType = type;
-                    bool foundMonoBehaviour = false;
-                    
-                    while (currentType != null && !foundMonoBehaviour)
-                    {
-                        if (currentType == typeof(MonoBehaviour))
-                        {
-                            foundMonoBehaviour = true;
-                        }
-                        currentType = currentType.BaseType;
-                    }
-                    
-                    Assert.IsTrue(foundMonoBehaviour, "LaserScanMsgPublisher should derive from MonoBehaviour");
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_DependencyChain_ShouldBeComplete()
-        {
-            // Test that all dependency classes in the chain are accessible
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // LaserScanMsgPublisher
-                var publisherType = System.Type.GetType("UnitySensors.ROS.Publisher.Sensor.LaserScanMsgPublisher, UnitySensorsROSRuntime");
-                Assert.IsNotNull(publisherType, "LaserScanMsgPublisher should be accessible");
-                
-                // RosMsgPublisher<T, TT>
-                var basePublisherType = System.Type.GetType("UnitySensors.ROS.Publisher.RosMsgPublisher`2, UnitySensorsROSRuntime");
-                Assert.IsNotNull(basePublisherType, "RosMsgPublisher base class should be accessible");
-                
-                // LaserScanMsgSerializer
-                var serializerType = System.Type.GetType("UnitySensors.ROS.Serializer.Sensor.LaserScanMsgSerializer, UnitySensorsROSRuntime");
-                Assert.IsNotNull(serializerType, "LaserScanMsgSerializer should be accessible");
-                
-                // RosMsgSerializer<T>
-                var baseSerializerType = System.Type.GetType("UnitySensors.ROS.Serializer.RosMsgSerializer`1, UnitySensorsROSRuntime");
-                Assert.IsNotNull(baseSerializerType, "RosMsgSerializer base class should be accessible");
-                
-                // HeaderSerializer
-                var headerSerializerType = System.Type.GetType("UnitySensors.ROS.Serializer.Std.HeaderSerializer, UnitySensorsROSRuntime");
-                Assert.IsNotNull(headerSerializerType, "HeaderSerializer should be accessible");
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_ROSMessageIntegration_ShouldWork()
-        {
-            // Test ROS message integration concepts for LaserScanMsgPublisher
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test LaserScanMsg structure concepts
-                var laserScanMsgProperties = new[] {
-                    "header", "angle_min", "angle_max", "angle_increment",
-                    "time_increment", "scan_time", "range_min", "range_max",
-                    "ranges", "intensities"
-                };
-                
-                // Test that all expected properties exist conceptually
-                foreach (var property in laserScanMsgProperties)
-                {
-                    Assert.IsNotNull(property);
-                    Assert.IsTrue(property.Length > 0);
-                }
-                
-                // Test ROS topic naming patterns for laser scan
-                var validLaserScanTopics = new[] {
-                    "/scan", "/laser_scan", "/sensor_msgs/laser_scan",
-                    "/lidar/scan", "/front_laser/scan", "/base_scan"
-                };
-                
-                foreach (var topic in validLaserScanTopics)
-                {
-                    Assert.IsNotNull(topic);
-                    Assert.IsTrue(topic.StartsWith("/"));
-                    Assert.IsTrue(topic.Contains("scan") || topic.Contains("laser") || topic.Contains("lidar"));
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_SensorDataFlow_ShouldWork()
-        {
-            // Test sensor data flow concepts for LaserScanMsgPublisher
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test data flow from point cloud to laser scan
-                var dataFlowSteps = new[] {
-                    "PointCloudGeneration", "PointCloudFiltering", "RangeCalculation",
-                    "NoiseApplication", "RangeFiltering", "IntensityMapping",
-                    "MessageSerialization", "ROSPublishing"
-                };
-                
-                // Test that all data flow steps are represented
-                foreach (var step in dataFlowSteps)
-                {
-                    Assert.IsNotNull(step);
-                    Assert.IsTrue(step.Length > 0);
-                }
-                
-                // Test point cloud to laser scan conversion concepts
-                var testPoint = new { x = 3.0f, y = 0.0f, z = 4.0f, intensity = 100.0f };
-                var expectedRange = System.Math.Sqrt(testPoint.x * testPoint.x + testPoint.z * testPoint.z);
-                
-                Assert.AreEqual(5.0f, expectedRange, 1e-6f);
-                Assert.IsTrue(testPoint.intensity >= 0.0f);
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_ConfigurationValidation_ShouldWork()
-        {
-            // Test configuration validation concepts for LaserScanMsgPublisher
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test typical laser scanner configuration ranges
-                var validConfigurations = new[] {
-                    new { minRange = 0.1f, maxRange = 10.0f, frequency = 10.0f, isValid = true },
-                    new { minRange = 0.5f, maxRange = 100.0f, frequency = 30.0f, isValid = true },
-                    new { minRange = 1.0f, maxRange = 200.0f, frequency = 60.0f, isValid = true },
-                    new { minRange = -1.0f, maxRange = 10.0f, frequency = 10.0f, isValid = false }, // Invalid min
-                    new { minRange = 5.0f, maxRange = 2.0f, frequency = 10.0f, isValid = false },   // Invalid range
-                    new { minRange = 0.1f, maxRange = 10.0f, frequency = -5.0f, isValid = false }   // Invalid frequency
-                };
-                
-                foreach (var config in validConfigurations)
-                {
-                    bool actualValid = config.minRange >= 0.0f && 
-                                     config.maxRange > config.minRange && 
-                                     config.frequency > 0.0f;
-                    
-                    Assert.AreEqual(config.isValid, actualValid);
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_PerformanceConsiderations_ShouldWork()
-        {
-            // Test performance considerations for LaserScanMsgPublisher
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test frequency and performance trade-offs
-                var performanceTests = new[] {
-                    new { frequency = 1.0f, pointCount = 360, complexity = "Low" },
-                    new { frequency = 10.0f, pointCount = 720, complexity = "Medium" },
-                    new { frequency = 30.0f, pointCount = 1440, complexity = "High" },
-                    new { frequency = 60.0f, pointCount = 2880, complexity = "VeryHigh" }
-                };
-                
-                foreach (var test in performanceTests)
-                {
-                    float computationLoad = test.frequency * test.pointCount;
-                    Assert.IsTrue(computationLoad > 0.0f);
-                    
-                    // Higher frequency and point count should mean higher complexity
-                    if (test.frequency >= 30.0f && test.pointCount >= 1440)
-                    {
-                        Assert.IsTrue(test.complexity == "High" || test.complexity == "VeryHigh");
-                    }
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_ErrorHandling_ShouldWork()
-        {
-            // Test error handling concepts for LaserScanMsgPublisher
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test error scenarios and handling
-                var errorScenarios = new[] {
-                    new { scenario = "NullPointCloudSource", shouldHandle = true },
-                    new { scenario = "InvalidScanPattern", shouldHandle = true },
-                    new { scenario = "ROSConnectionFailure", shouldHandle = true },
-                    new { scenario = "SerializationError", shouldHandle = true },
-                    new { scenario = "InvalidConfiguration", shouldHandle = true }
-                };
-                
-                foreach (var error in errorScenarios)
-                {
-                    Assert.IsNotNull(error.scenario);
-                    Assert.IsTrue(error.shouldHandle);
-                    
-                    // Test that error handling is considered
-                    bool hasErrorHandlingConcepts = error.scenario.Contains("Null") || 
-                                                   error.scenario.Contains("Invalid") || 
-                                                   error.scenario.Contains("Failure") || 
-                                                   error.scenario.Contains("Error");
-                    
-                    Assert.IsTrue(hasErrorHandlingConcepts);
-                }
-            });
-        }
-
-        [Test]
-        public void LaserScanMsgPublisher_IntegrationComplete_ShouldWork()
-        {
-            // Test complete integration of LaserScanMsgPublisher system
-            // Act & Assert
-            Assert.DoesNotThrow(() => {
-                // Test end-to-end integration concepts
-                var integrationComponents = new[] {
-                    "UnityMonoBehaviour", "ROSConnection", "MessageSerialization",
-                    "PointCloudProcessing", "TimingControl", "ConfigurationManagement",
-                    "ErrorHandling", "PerformanceOptimization"
-                };
-                
-                // Verify all integration components are considered
-                foreach (var component in integrationComponents)
-                {
-                    Assert.IsNotNull(component);
-                    Assert.IsTrue(component.Length > 0);
-                }
-                
-                // Test system lifecycle
-                var lifecycleStates = new[] {
-                    "Initialization", "Configuration", "Activation", 
-                    "Publishing", "Monitoring", "Cleanup"
-                };
-                
-                foreach (var state in lifecycleStates)
-                {
-                    Assert.IsNotNull(state);
-                    Assert.IsTrue(state.Length > 0);
-                }
-                
-                // Final integration test - verify all major components work together
-                var systemIntegrationComplete = true;
-                Assert.IsTrue(systemIntegrationComplete, "LaserScanMsgPublisher integration should be complete");
+                // Verify that the test suite is complete
+                var testSuiteComplete = true;
+                Assert.IsTrue(testSuiteComplete, "UnitySensors test suite should be complete and organized");
             });
         }
     }
