+++ Assets/UnitySensors/Tests/Editor/Vector3DBasicTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d683ba7c487cde26d9685f96afc3f4a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
